% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_param-class.R
\name{ALL}
\alias{ALL}
\title{PARAM object class constructor}
\usage{
ALL()
}
\value{
ALL param object
}
\description{
This class constructor is used to create instances of PARAM object
to be used in GMQL functions
It defines the minimum (and/or maximum) to the number of samples 
in the input dataset;
}
\examples{

## This statement produces an output dataset with a single output sample. 
## The COVER operation considers all areas defined by a minimum 
## of two overlapping regions in the input samples, 
## up to maximum amount of overlapping regions.

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)
res = cover(exp, 2, "ALL")

}
\seealso{
\code{\link{ANY}}
}
