% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_param-class.R
\name{ANY}
\alias{ANY}
\title{PARAM object class constructor}
\usage{
ANY()
}
\value{
ANY param object
}
\description{
This class constructor is used to create instances of PARAM object
to be used in GMQL functions
It defines any amount of overlapping regions to be considered.
}
\examples{

## This statement produces an output dataset with a single output sample. 
## The COVER operation considers all areas defined by a minimum 
## of two overlapping regions in the input samples, 
## up to any amount of overlapping regions.

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)
res = cover(input_data = exp, 2, ANY())


}
\seealso{
\code{\link{ALL}}
}
