% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{AVG}
\alias{AVG}
\title{OPERATOR object class constructor}
\usage{
AVG(value)
}
\arguments{
\item{value}{single string identifying name of metadata attribute}
}
\value{
AVG Operator object
}
\description{
This class constructor is used to create instances of OPERATOR object
to be used in GMQL functions that require operator on value,
it prepared input parameter to be passed to library function arithmetic mean,
performing all the type conversion needed
}
\examples{

### local with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
exp = readDataset(test_path)

## The following cover operation produces output regions where at least 2 and at most 3 regions of
## exp overlap, having as resulting region attributes the avg signal of the overlapping regions;
## the result has one sample for each input cell.
res = cover(input_data = exp,2,3, c("cell"), list(avg_signal = AVG("signal")))

}
\seealso{
\code{\link{MIN}} \code{\link{MAX}} \code{\link{SUM}} \code{\link{COUNT}}
\code{\link{BAG}} \code{\link{STD}} \code{\link{MEDIAN}} \code{\link{Q1}} \code{\link{Q2}}
\code{\link{Q3}}
}
