% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{BAG}
\alias{BAG}
\title{OPERATOR object class constructor}
\usage{
BAG(value)
}
\arguments{
\item{value}{single string identifying name of region attribute}
}
\value{
BAG Operator object
}
\description{
This class constructor is used to create instances of OPERATOR object
to be used in GMQL functions that require operator on value,
it prepared input parameter to be passed to library function bag,
this function creates comma-separated strings of distinct attribute values
performing all the type conversion needed
}
\examples{

### local with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
data = readDataset(test_path)

### copies all samples of DATA into OUT dataset, and then for each of them adds another 
metadata attribute,  allScores, which is the aggregation comma-separated list of all the 
distinct values that the attribute  score takes in the sample.
out = extend(input_data = data, list(allScore = BAG("score"))

}
\seealso{
\code{\link{MIN}} \code{\link{MAX}} \code{\link{AVG}} \code{\link{COUNT}}
\code{\link{SUM}} \code{\link{STD}} \code{\link{MEDIAN}} \code{\link{Q1}} \code{\link{Q2}}
\code{\link{Q3}}
}
