% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-class.R
\name{BAGD}
\alias{BAGD}
\title{AGGREGATES object class constructor}
\usage{
BAGD(value)
}
\arguments{
\item{value}{string identifying name of region attribute}
}
\value{
BAGD aggregate object
}
\description{
This class constructor is used to create instances of AGGREGATES object,
to be used in GMQL functions that require aggregate on value.
It prepares input parameter to be passed to the library function bagd,
this function creates comma-separated strings of distinct attribute values
performing all the type conversions needed
}
\examples{

## local with CustomParser
init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
data = read_dataset(test_path)

## It copies all samples of DATA into OUT dataset, and then for each of 
## them it adds another metadata attribute, allScores, which is the 
## aggregation comma-separated list of all the distinct values that 
## the region attribute score takes in the sample.

out = extend(data, allScore = BAGD("score"))

}
\seealso{
\code{\link{SUM}} \code{\link{COUNT}} \code{\link{MIN}} 
\code{\link{MAX}} \code{\link{AVG}} \code{\link{MEDIAN}} 
\code{\link{STD}} \code{\link{BAG}}
\code{\link{Q1}} \code{\link{Q2}} \code{\link{Q3}}
}
