% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gmql_materialize.R
\docType{methods}
\name{collect}
\alias{collect}
\alias{collect,GMQLDataset-method}
\alias{collect-method}
\title{Method collect}
\usage{
collect(x, dir_out = getwd(), name = "ds1", ...)

\S4method{collect}{GMQLDataset}(x, dir_out = getwd(), name = "ds1")
}
\arguments{
\item{x}{GMQLDataset class object}

\item{dir_out}{destination folder path.
by default is current working directory of the R process}

\item{name}{name of the result dataset}

\item{...}{Additional arguments for use in specific methods}
}
\value{
None
}
\description{
Wrapper to GMQL materialize function

It saves the contents of a dataset that contains samples metadata and 
samples regions.
It is normally used to persist the contents of any dataset generated 
during a GMQL query.
Any dataset can be materialized, but the operation can be time-consuming.
For best performance, materialize the relevant data only.
}
\examples{

init_gmql()
test_path <- system.file("example","DATASET",package = "RGMQL")
r = read_dataset(test_path)
s = filter(r)
m = aggregate(s, DF("antibody_targer","cell_karyotype"))
collect(m, dir_out = test_path)

}
