% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{compile_query}
\alias{compile_query}
\alias{compile_query}
\alias{compile_query_fromfile}
\title{Compile GMQL query}
\usage{
compile_query(url, query)

compile_query_fromfile(url, filePath)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{query}{string text of the query}

\item{filePath}{string path of txt files containing a GMQL query}
}
\value{
None
}
\description{
It compiles the GMQL query into repository taken from file or insert as text 
string, using the proper GMQL web service available on a remote server
}
\examples{

remote_url = "http://genomic.elet.polimi.it/gmql-rest-r/"
login_gmql(remote_url)
compile_query(remote_url, "DATASET = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATASET INTO RESULT_DS;")

\dontrun{
test_path <- system.file("example", package = "RGMQL")
test_query <- file.path(test_path, "query1.txt")
compile_query_fromfile(remote_url, test_query)
}

}
