% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{compile_query_fromfile}
\alias{compile_query_fromfile}
\title{Compile GMQL query from file}
\usage{
compile_query_fromfile(url, filePath)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{filePath}{string path of txt files containing a GMQL query}
}
\value{
None
}
\description{
It compiles a GMQL query from file using the proper GMQL web service 
available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
test_path <- system.file("example", package = "RGMQL")
test_query <- file.path(test_path, "query1.txt")
compile_query_fromfile(remote_url, test_query)

}
