% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{compileQuery.fromfile}
\alias{compileQuery.fromfile}
\title{Compile GMQL query from file}
\usage{
compileQuery.fromfile(url, filePath)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{filePath}{single string path of txt files containing a GMQL query}
}
\value{
None
}
\description{
It compiles a GMQL query from file using the proper GMQL web service available 
on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl)
test_path <- system.file("example",package = "GMQL")
test_query <- file.path(test_path, "query1.txt")
compileQuery.fromfile(PolimiUrl,test_query)

}
