% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-functions.R
\name{Evaluation-Function}
\alias{Evaluation-Function}
\alias{conds}
\alias{condition_evaluation}
\title{Condition evaluation functions}
\usage{
conds(default = c(""), full = c(""), exact = c(""))
}
\arguments{
\item{default}{concatenation of string identifying a name of metadata 
attribute to be evaluated. 
It defines a DEFAULT evaluation of the input values. 
DEFAULT evaluation: the two attributes match if both end with value.}

\item{full}{concatenation of string identifying a name of metadata 
attribute to be evaluated.
It defines a FULL (FULLNAME) evaluation of the input values.
FULL evaluation: two attributes match if they both end with value and,
if they have further prefixes, the two prefix sequences are identical.}

\item{exact}{concatenation of string identifying a name of metadata 
attribute to be evaluated.
It defines a EXACT evaluation of the input values.
EXACT evaluation: only attributes exactly as value match; 
no further prefixes are allowed.}
}
\value{
list of 2-D array containing method of evaluation and metadata 
attribute name
}
\description{
This function is used to support joinBy and/or groupBy function parameter.
}
\examples{

"where is my example?"

}
