% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{COUNT}
\alias{COUNT}
\title{OPERATOR object class constructor}
\usage{
COUNT()
}
\value{
COUNT Operator object
}
\description{
This class constructor is used to create instances of OPERATOR object
to be used in GMQL functions that require operator,
it prepared input parameter to be passed to library function count,
performing all the type conversion needed
}
\examples{

### local with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
exp = readDataset(test_path)

### counts the regions in each sample and stores their number as value of the new metadata 
RegionCount attribute of the sample.
out = extend(input_data = exp, list(RegionCount = COUNT())

}
\seealso{
\code{\link{MIN}} \code{\link{MAX}} \code{\link{AVG}} \code{\link{SUM}}
\code{\link{BAG}} \code{\link{STD}} \code{\link{MEDIAN}} \code{\link{Q1}} \code{\link{Q2}}
\code{\link{Q3}}
}
