% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_param-class.R
\name{COVER-PARAMETER}
\alias{COVER-PARAMETER}
\alias{ALL}
\alias{COVER-PARAMETER}
\alias{ANY}
\title{PARAM object class constructor}
\usage{
ALL()

ANY()
}
\value{
param object
}
\description{
This class constructor is used to create instances of PARAM object
to be used in GMQL cover method
}
\details{
\itemize{
\item{ANY: It defines any amount of overlapping regions to be considered.}
\item{ALL: It defines the minimum (and/or maximum) to the number of samples 
in the input dataset.}
}
}
\examples{

#' init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)

## This statement produces an output dataset with a single output sample. 
## The COVER operation considers all areas defined by a minimum 
## of two overlapping regions in the input samples, 
## up to maximum amount of overlapping regions.

res = cover(exp, 2, "ALL")

## This statement produces an output dataset with a single output sample. 
## The COVER operation considers all areas defined by a minimum 
## of two overlapping regions in the input samples, 
## up to any amount of overlapping regions.

res = cover(exp, 2, "ANY")

}
