% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-class.R
\name{DEF}
\alias{DEF}
\title{CONDITION object class constructor}
\usage{
DEF(value)
}
\arguments{
\item{value}{single string identifying name of metadata and/or region attribute
to be evaluated}
}
\value{
DEF condition object
}
\description{
This class constructor is used to create instances of CONDITION object
to be used in GMQL functions that require evaluation on value
This constructor define a DEF (DEFAULT) evaluation of the input value.
DEF evaluation: two attributes match if both end with value.
}
\examples{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "RGMQL")
r = readDataset(test_path)

#### select with DEF condition
s = select(input_data = r, semi_join = c("cell_type","cell"),
semi_join_dataset = c)

\dontrun{

#### select with DEF condition declared explicitily
s = select(input_data = r, semi_join = c(DEF("attribute_tag"),"cell_type","cell"),
semi_join_dataset = c)

}

}
\seealso{
\code{\link{FULL}} \code{\link{EXACT}}
}
