% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Delete dataset}
\usage{
deleteDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{datasetName}{single string name of dataset to delete}
}
\value{
None
}
\description{
It deletes single private dataset specified by name from repository using the proper GMQL 
web service available on a remote server
}
\details{
If no error occures print "Deleted Dataset", otherwise a specific error is printed
}
\examples{

\dontrun{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl,"test101","test")
deleteDataset(PolimiUrl,"job_test1_test101_20170604_180908_RESULT_DS")
}

}
\seealso{
\code{\link{downloadDataset}}
}
