% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DG}
\alias{DG}
\title{DISTAL object class constructor}
\usage{
DG(value)
}
\arguments{
\item{value}{single string identifying distance between genomic regions 
in base pairs}
}
\value{
None
}
\description{
This class constructor is used to create instances of DISTAL object
to be used in GMQL functions that use genometric predicate parameter
requiring distal condition on value
In this case DLE: denotes the great distance clause, which selects all 
the regions of the experiment such that their distance from the anchor 
region is less than, or equal to, 'value' bases.
}
\examples{

## Given a dataset HM and one called TSS with a sample including 
# Transcription Start Site annotations, it searches for those regions of hm 
# that are at a minimal distance from a transcription start site 'TSS' 
# and takes the first/closest one for each TSS, provided that such distance 
# is greater than 12 bases and joined TSS and HM samples are obtained 
# from the same provider 'joinby clause'.

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
test_path2 <- system.file("example", "DATASET_GDM", package = "RGMQL")
TSS = read_dataset(test_path)
HM = read_dataset(test_path2)
join_data = join(TSS, HM, 
genometric_predicate = list(list(MD(1), DG(12))), c("provider"), 
region_output = "RIGHT")


}
\seealso{
\code{\link{DGE}} \code{\link{DLE}} \code{\link{DL}} 
\code{\link{MD}} \code{\link{DOWN}} \code{\link{UP}}
}
