% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DLE}
\alias{DLE}
\title{DISTAL object class constructor}
\usage{
DLE(value)
}
\arguments{
\item{value}{single string identifying distance between genomic regions in base pairs,}
}
\value{
None
}
\examples{
\dontrun{

}
""
}
\seealso{
\code{\link{DGE}} \code{\link{MD}}  \code{\link{DOWN}} \code{\link{UP}}
}
