% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{download_as_GRangesList}
\alias{download_as_GRangesList}
\title{Download Dataset in GrangesList}
\usage{
download_as_GRangesList(url, datasetName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{datasetName}{string name of dataset we want to get}
}
\value{
GrangesList containing all GMQL sample in dataset
}
\description{
It donwloads private dataset from repository saving into R environemnt 
as GrangesList
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{
## create grangeslist from public dataset HG19_BED_ANNOTATION got 
## from repository

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
download_as_GRangesList(remote_url, "public.HG19_BED_ANNOTATION")
}

}
