% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{download_dataset}
\alias{download_dataset}
\alias{download_as_GRangesList}
\title{Download Dataset}
\usage{
download_dataset(url, datasetName, path = getwd())

download_as_GRangesList(url, datasetName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name will be added automatically}

\item{datasetName}{string name of dataset we want to get}

\item{path}{string local path folder where store dataset,
by defualt is R working directory}
}
\value{
None

GrangesList containing all GMQL sample in dataset
}
\description{
It donwloads private dataset as zip file from repository to local path or
donwloads saving it into R environemnt as GrangesList using the proper 
GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

## download dataset in r working directory
## in this case we try to download public dataset

\dontrun{

remote_url = "http://genomic.elet.polimi.it/gmql-rest-r"
login_gmql(remote_url)
download_dataset(remote_url, "public.HG19_BED_ANNOTATION", path = getwd())

## create grangeslist from public dataset HG19_BED_ANNOTATION got 
## from repository

download_as_GRangesList(remote_url, "public.HG19_BED_ANNOTATION")
}

}
