% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{downloadDatasetToGrangesList}
\alias{downloadDatasetToGrangesList}
\title{Download Dataset in GrangesList}
\usage{
downloadDatasetToGrangesList(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{datasetName}{single string name of dataset we want to get}
}
\value{
GrangesList containing all GMQL sample in dataset
}
\description{
It donwloads private dataset from repository saving into R environemnt as GrangesList
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{
#### create grangeslist from public dataset HG19_BED_ANNOTATION got from repository
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(url = PolimiUrl)
downloadDatasetToGrangesList(PolimiUrl,"public.HG19_BED_ANNOTATION")
}

}
