% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-class.R
\name{EXACT}
\alias{EXACT}
\title{CONDITION object class constructor}
\usage{
EXACT(value)
}
\arguments{
\item{value}{string identifying name of metadata attribute 
to be evaluated}
}
\value{
EXACT condition object
}
\description{
This class constructor is used to create instances of CONDITION object,
to be used in GMQL functions that require evaluation on value.
This constructor defines a EXACT evaluation of the input value.
EXACT evaluation: only attributes exactly as value will match; 
no further prefixes are allowed.
}
\examples{

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
r = read_dataset(test_path)

## select with condition
## the first and the third attribute are DEF the second one is EXACT
s = filter(r, semi_join = list("cell_type", EXACT("cell"), 
"attribute_tag"), semi_join_dataset = r)

\dontrun{

## select with default condition
## the EXACT condition is treated as default due to coercion
s = filter(r, semi_join = c("cell_type", "cell", EXACT("attribute_tag")),
semi_join_dataset = r)

}

}
\seealso{
\code{\link{FULL}}
}
