% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesToGMQL.R
\name{export_gmql}
\alias{export_gmql}
\title{Create GMQL dataset from GrangesList}
\usage{
export_gmql(samples, dir_out, is_gtf)
}
\arguments{
\item{samples}{GrangesList}

\item{dir_out}{folder path where create a folder and write the sample files}

\item{is_gtf}{logical value indicating if samples have to be exported
with GTF or GDM format}
}
\value{
None
}
\description{
It create GMQL dataset from GRangesList.
All samples are in GDM (tab-separated values) or GTF file format
}
\details{
The GMQL dataset is made up by two differet file type

\itemize{
\item{metadata files: contain metadata associated to corrisponding sample}
\item{region files: contain many genomic regions }
\item{region schema file: XML file contains region attributes 
(e.g. chr, start, end, pvalue)}
}
region sample files and metadata files are associated through file name:
for example S_0001.gdm for region file and S_0001.gdm.meta for 
its metadata file
}
\examples{

\dontrun{
library(GenomicRanges)
gr1 <- GRanges(seqnames = "chr2", ranges = IRanges(3, 6), strand = "+", 
score = 5L, GC = 0.45)
gr2 <- GRanges(seqnames = c("chr1", "chr1"),
ranges = IRanges(c(7,13), width = 3), strand = c("+", "-"), 
score = 3:4, GC = c(0.3, 0.5))
grl = GRangesList(gr1, gr2)
test_out_path <- system.file("example", package = "RGMQL")
export_gmql(grl, test_out_path,TRUE)
}
}
\seealso{
\code{\link{import_gmql}}
}
