% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesToGMQL.R
\name{exportGMQL.gtf}
\alias{exportGMQL.gtf}
\title{Create GMQL dataset from GrangesList}
\usage{
exportGMQL.gtf(samples, dir_out)
}
\arguments{
\item{samples}{GrangesList}

\item{dir_out}{folder path where create a folder and write all the sample files}
}
\value{
None
}
\description{
It create GMQL dataset from GRangesList.
All samples are in GTF file format
}
\details{
The GMQL dataset is made up by two differet file type
\itemize{
\item{metadata files: contain metadata associated to corrisponding sample}
\item{region files: contains many genomic regions }
\item{region schema file: XML file contains region attributes (e.g. chr, left, right, qvalue)}
}
region sample files and metadata files are associated through file name:
for example S_0001.gtf for region file and S_0001.gtf.meta for its metadata file


}
\examples{

library(GenomicRanges)
gr1 <- GRanges(seqnames = "chr2", ranges = IRanges(3, 6), strand = "+", score = 5L, GC = 0.45)
gr2 <- GRanges(seqnames = c("chr1", "chr1"),
ranges = IRanges(c(7,13), width = 3), strand = c("+", "-"), score = 3:4, GC = c(0.3, 0.5))
grl = GRangesList(gr1,gr2)
test_out_path <- system.file("example",package = "RGMQL")
exportGMQL.gtf(grl,test_out_path)

}
\seealso{
\code{\link{exportGMQL.gdm}} \code{\link{exportGMQL.gtf}} \code{\link{importGMQL.gdm} }
}
