% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Cover-Param.R
\name{filter-extract}
\alias{filter-extract}
\alias{FULL}
\title{PARAM object class constructor}
\usage{
FULL(except = NULL)
}
\arguments{
\item{except}{The list of attribute to not consider}
}
\value{
Param object
}
\description{
This class constructor is used to create instances of PARAM object
to be used in filter and extract function.
}
\details{
It is used to encompasses all the region parameters already present 
into the dataset or GrangesList

\itemize{
\item{FULL: It consider all the region paramter}
}
}
