% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-class.R
\name{FULL}
\alias{FULL}
\title{CONDITION object class constructor}
\usage{
FULL(value)
}
\arguments{
\item{value}{single string identifying name of metadata and/or region attribute
to be evaluated}
}
\value{
FULL condition object
}
\description{
This class constructor is used to create instances of CONDITION object
to be used in GMQL functions that require evaluation on value
This constructor define a FULL (FULLNAME) evaluation of the input value.
FULL evaluation: two attributes match if they both end with value and,
if they have further prefixes, the two prefix sequences are identical
}
\examples{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = readDataset(test_path)

#### select with condition
#### the first and the third attribute are DEF the second one is FULL
s = select(input_data = r, semi_join = list("cell_type",FULL("cell"),attribute_tag), semi_join_dataset = c)

\dontrun{

#### select with DEF condition
#### the FULL condition is treated as DEF due to coercion
s = select(input_data = r, semi_join = c("cell_type","cell",FULL("attribute_tag")),
semi_join_dataset = c)

#### select with condition
#### the first is FULL and the other ones are DEF
s = select(input_data = r, semi_join = c(FULL("attribute_tag"),"cell_type","cell"),
semi_join_dataset = c)

}

}
