% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gmql_group.R
\docType{methods}
\name{group_by}
\alias{group_by}
\alias{group_by,GMQLDataset-method}
\alias{group_by}
\alias{group_by-method}
\title{Method group_by}
\usage{
group_by(.data, ...)

\S4method{group_by}{GMQLDataset}(.data, ...)
}
\arguments{
\item{.data}{GMQLDataset object}
}
\value{
GMQLDataset object. It contains the value to use as input 
for the subsequent GMQLDataset method
}
\description{
Wrapper to GMQL group function

Method group_by
}
\examples{



}
