% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{import_gmql}
\alias{import_gmql}
\title{Create GRangesList from GMQL dataset}
\usage{
import_gmql(dataset_path, is_gtf)
}
\arguments{
\item{dataset_path}{string with GMQL dataset folder path}

\item{is_gtf}{logical value indicating if dataset samples are in GTF format;
if TRUE and dataset does not contain GTF samples an error occurs}
}
\value{
GRangesList containing all GMQL samples in dataset
}
\description{
It creates a GRangesList from GMQL samples in dataset. 
It reads sample files in GTF or GDM/tab-delimited format.
}
\examples{

## This statement defines the path to the subdirectory "example" of the 
## package "RGMQL" and imports as GRangesList the contained GMQL dataset

test_path <- system.file("example", "DATASET", package = "RGMQL")
grl = import_gmql(test_path, TRUE)


}
\seealso{
\code{\link{export_gmql}}
}
