% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{importGMQL.gdm}
\alias{importGMQL.gdm}
\title{Create GrangesList from dataset GMQL in GDM (delimited / tabulated) format file}
\usage{
importGMQL.gdm(datasetName)
}
\arguments{
\item{datasetName}{single string GMQL dataset folder path}
}
\value{
GrangesList containing all GMQL sample In dataset
}
\description{
It create a GrangesList from GMQL samples in dataset
It reads only sample files in GDM format
}
\examples{

test_path <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
grl = importGMQL.gdm(test_path)

}
\seealso{
\code{\link{exportGMQL.gdm}} \code{\link{exportGMQL.gtf}} \code{\link{importGMQL.gtf} }
}
