% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{init_gmql}
\alias{init_gmql}
\title{Init GMQL Server}
\usage{
init_gmql(output_format = "gtf", remote_processing = FALSE, url = NULL,
    username = NULL, password = NULL)
}
\arguments{
\item{output_format}{string identifies the output format of sample files.
Can be TAB, GTF or COLLECT
\itemize{
\item{TAB: tab delimited file format}
\item{GTF: file format used to hold information about gene structure.
It is a tab-delimited text format based on the general feature format}
\item{COLLECT: used for storing output in memory}
}}

\item{remote_processing}{logical value specifying the processing mode.
True for processing on cluster (remote), false for local processing.}

\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically.
If null, no login is performed.
You can always perform it, calling the function \code{\link{login_gmql}} 
explicitly}

\item{username}{string name used during signup}

\item{password}{string password used during signup}
}
\value{
None
}
\description{
Initialize and run GMQL server for executing GMQL query
It is also perform a login to GMQL REST services suite if needed
}
\examples{

## initialize GMQL with local processing with sample files output format 
## as Tab delimited
library(rJava)
init_gmql("tab", FALSE)

\dontrun{

## initialize GMQL with remote processing
remote_url = "http://130.186.13.219/gmql-rest"
init_gmql(remote_processing = TRUE, url = remote_url)

}

}
