% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{initGMQL}
\alias{initGMQL}
\title{Init GMQL Server}
\usage{
initGMQL(output_format = "gtf", remote_processing = FALSE, url = NULL,
  username = NULL, password = NULL)
}
\arguments{
\item{output_format}{single string identifies the output format of sample files.
Can be TAB, GTF or COLLECT
\itemize{
\item{TAB: tab delimited file format}
\item{GTF: file format used to hold information about gene structure.
It is a tab-delimited text format based on the general feature format}
\item{COLLECT: used for storing output in memory}
}}

\item{remote_processing}{logical value specifying the processing mode.
True for processing on cluster (remote), false for local processing (local machine)}

\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically.
If null no login is performed, you can always perform it calling the function 
\code{\link{login.GMQL}} explicitly}

\item{username}{single string name used during signup}

\item{password}{single string password used during signup}
}
\value{
None
}
\description{
Initialize and run GMQL server for executing GMQL query
It is also perform a login to GMQL REST services suite if needed
}
\examples{

# initialize GMQL with local processing
library(rscala)
initGMQL("tab",FALSE)

}
