% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{log_job}
\alias{log_job}
\alias{show_job_log}
\alias{trace_job}
\title{Show a job log or trace}
\usage{
show_job_log(url, job_id)

trace_job(url, job_id)
}
\arguments{
\item{url}{string url of server: It must contain the server address
and base url; service name is added automatically}

\item{job_id}{string id of the job}
}
\value{
log or trace text
}
\description{
It show a job log or traces a specific job
}
\details{
If error occures a specific error is printed
}
\examples{

remote_url = "http://genomic.elet.polimi.it/gmql-rest-r"
login_gmql(remote_url)

## list all jobs
list_jobs <- show_jobs_list(remote_url)

\dontrun{
jobs_1 <- list_jobs$jobs[[1]]

## show job log
show_job_log(remote_url, jobs_1)

## trace job
trace_job(remote_url, jobs_1)

}

}
