% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOp.R
\name{logout.GMQL}
\alias{logout.GMQL}
\title{Logout from GMQL}
\usage{
logout.GMQL(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}
}
\value{
None
}
\description{
Logout from GMQL REST services suite
using the proper GMQL web service available on a remote server
}
\details{
After logout the authentication token will be invalidated.
The authentication token is removed from Global environment
If error occures a specific error is printed
}
\examples{

#### login as guest, then logout
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl)
logout.GMQL(PolimiUrl)

\dontrun{
##### login with username and password, then logout
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl,"test101","test")
logout.GMQL(PolimiUrl)
}

}
\seealso{
\code{\link{register.GMQL}} \code{\link{login.GMQL}}
}
