% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOp.R
\name{logout_gmql}
\alias{logout_gmql}
\title{Logout from GMQL}
\usage{
logout_gmql(url)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
None
}
\description{
Logout from GMQL REST services suite
using the proper GMQL web service available on a remote server
}
\details{
After logout the authentication token will be invalidated.
The authentication token is removed from Global environment
If error occures a specific error is printed
}
\examples{
#' ## login as guest, then logout
remote_url = "http://130.186.13.219/gmql-rest"
\dontrun{
login_gmql(remote_url)
logout_gmql(remote_url)
}
}
\seealso{
\code{\link{login_gmql}}
}
