% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLDataset-class.R, R/Materialize.R
\docType{methods}
\name{materialize}
\alias{materialize}
\alias{materialize,GMQLDataset-method}
\alias{materialize-method}
\title{Method materialize}
\usage{
materialize(data, ...)

\S4method{materialize}{GMQLDataset}(data, dir_out = getwd())
}
\arguments{
\item{data}{GMQLDataset class object}

\item{...}{Additional arguments for use in specific methods}

\item{dir_out}{destination folder path.
by default is current working directory of the R process}
}
\value{
None
}
\description{
Wrapper to GMQL materialize function

It saves the contents of a dataset that contains samples metadata and 
samples regions.
It is normally used to persist the contents of any dataset generated 
during a GMQL query.
Any dataset can be materialized, but the operation can be time-consuming.
For best performance, materialize the relevant data only.
}
\examples{

init_gmql()
test_path <- system.file("example","DATASET",package = "RGMQL")
r = read_dataset(test_path)
s = filter(r)
m = aggregate(s, DF("antibody_targer","cell_karyotype"))
materialize(m, dir_out = test_path)

}
