% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{MD}
\alias{MD}
\title{DISTAL object class constructor}
\usage{
MD(value)
}
\arguments{
\item{value}{single string identifying number of regions at minimum distance beetwen expertiment
and anchor region}
}
\value{
None
}
\description{
This class constructor is used to create instances of DISTAL object
to be used in GMQL functions \code{\link{join}} in genometric predicate parameter
that require distal condition on value
In this case MD: denotes the minimum distance clause, which selects the 'value' regions of the experiment 
at minimial distance from the anchor region.
When there are ties (i.e., regions at the same distance from the anchor region), 
regions of the experiment are kept in the result even if they exceed the 'value' limit.
}
\examples{
\dontrun{

}
""
}
\seealso{
\code{\link{DLE}} \code{\link{DGE}}  \code{\link{DOWN}} \code{\link{UP}}
}
