% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{metadataFromSample}
\alias{metadataFromSample}
\title{Shows metadata list from dataset sample}
\usage{
metadataFromSample(url, datasetName, sampleName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{datasetName}{single string name of dataset to get}

\item{sampleName}{single string sample name to get}
}
\value{
list of metadata in the form 'key = value'
}
\description{
It retrieves metadata for a specific sample in dataset using the proper GMQL web service 
available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

## download metadata with real test login
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(url = PolimiUrl)
metadataFromSample(PolimiUrl,"public.HG19_BED_ANNOTATION","genes")

}
