% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{NIL}
\alias{NIL}
\title{OPERATOR object class constructor}
\usage{
NIL(value)
}
\arguments{
\item{value}{string identifying name of region attribute}
}
\value{
NIL operator object
}
\description{
This class constructor is used to create instances of OPERATOR object,
to be used in GMQL functions that require operator on value.
It prepared input parameter to be passed to library function nil,
performing all the type conversion needed
}
\examples{

## It define a new numeric region attribute with "null" value. 
## The syntax for creating a new attribute with null value is 
## attribute_name = NULL(TYPE), where type may be INTEGER or DOUBLE.

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)
out = subset(exp, regions_update = list(signal = NIL("INTEGER"), 
pvalue = NIL("DOUBLE")))


}
\seealso{
\code{\link{META}} \code{\link{SQRT}}
}
