% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-functions.R
\name{Ordering-functions}
\alias{Ordering-functions}
\alias{DESC}
\alias{Ordering-functions}
\alias{ASC}
\title{Ordering functions}
\usage{
DESC(...)

ASC(...)
}
\arguments{
\item{...}{series of metatdata as string}
}
\value{
Ordering object
}
\description{
These functions is used to create a series of metadata as string
that require ordering on value; is used only in arrange method
(see example).
}
\details{
\itemize{
\item{ASC: It defines an ascending order for input value}
\item{DESC: It defines a descending order for input value}
}
}
\examples{

## This statement initializes and runs the GMQL server for local execution 
## and creation of results on disk. Then, with system.file() it defines 
## the path to the folder "DATASET" in the subdirectory "example"
## of the package "RGMQL" and opens such file as a GMQL dataset named 
## "data" using customParser

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
data = read_dataset(test_path)

## It orders the samples according to the Region_Count metadata attribute 
## and takes the two samples that have the lowest count. 

asc = arrange(data, list(ASC("Region_Count")), fetch_opt = "mtop", 
num_fetch = 2)

## It orders the samples according to the pvalue schema attribute 
## and takes the first five samples that have the lowest pvalue and the 
## first seven regions.
 
desc = arrange(data, list(DESC("pvalue")), fetch_opt = "mtop", num_fetch = 5, 
reg_fetch_opt = "rtop", reg_num_fetch = 7)

}
