% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(samples)
}
\arguments{
\item{samples}{GrangesList}
}
\value{
DataSet class object. It contains the value to use as input 
for the subsequent GMQL function
}
\description{
Read a GrangesList saving in scala memory that can be referenced in R
}
\examples{

library("GenomicRanges")
gr1 <- GRanges(seqnames = "chr2", ranges = IRanges(103, 106),
strand = "+", score = 5L, GC = 0.45)
gr2 <- GRanges(seqnames = c("chr1", "chr1"), ranges = IRanges(c(107, 113), 
width = 3), strand = c("+", "-"), score = 3:4, GC = c(0.3, 0.5))

grl <- GRangesList("txA" = gr1, "txB" = gr2)

data_out <- read(grl)

}
