% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{regionFromSample}
\alias{regionFromSample}
\title{Shows regions from a dataset sample}
\usage{
regionFromSample(url, datasetName, sampleName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{datasetName}{single string name of dataset to get}

\item{sampleName}{single string sample name to get}
}
\value{
Granges data containing regions of sample
}
\description{
It retrieves regions for a specific sample (whose name is specified in the paramter "sampleName")
in a specific dataset (whose name is specified in the paramter "datasetName") 
using the proper GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl,"test101","test")
regionFromSample(PolimiUrl,"job_test1_test101_20170604_180908_RESULT_DS","S_00000")

}
