% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOp.R
\name{register.GMQL}
\alias{register.GMQL}
\title{Sign up to GMQL}
\usage{
register.GMQL(url, name, lastname, mail, username, password)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{name}{single string name of the user (can contain space)}

\item{lastname}{single string last name of the user (can contain space)}

\item{mail}{single string email (as spacified in RFC format)}

\item{username}{single string name you want to authenticate with}

\item{password}{single string password you want to authenticate with}
}
\value{
None
}
\description{
Sign up to GMQL REST services suite
using the proper GMQL web service available on a remote server
}
\details{
After registration you receive an authentication token, (i.e you are logged in after sing up,
no need to subsequent calling of login function)
As token remains vaild on server (until the next login / registration) a user can safely use a token for
a previous session as a convenience, this token is saved in Global environment to perform
subsequent REST call even on complete R restart (if is environemnt has been saved, of course ...)
If error occures a specific error is printed
}
\examples{

##### this user already exist, it's a test account
##### don't use it
PolimiUrl = "http://130.186.13.219/gmql-rest"
register.GMQL(url = PolimiUrl,"jonh","Doe","jonh@doe.com","JD","JD46")

}
\seealso{
\code{\link{login.GMQL}} \code{\link{logout.GMQL}}
}
