% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{remote_processing}
\alias{remote_processing}
\title{Disable or Enable remote processing}
\usage{
remote_processing(is_remote)
}
\arguments{
\item{is_remote}{sinble logical value, TRUE you will set a remote query processing mode 
otherwise will be local}
}
\description{
It allows to enable or disable remote processing
}
\details{
The invocation of this function allow to change mode of processing.
Thw switch is posible at the beginning when you didn't run any query at all, or after an execution 
(or take) function
}
\examples{

# initialize with remote processing off
initGMQL("tab",remote_processing=FALSE)

# change processing mode to remote
remote_processing(TRUE)

}
