% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{run_query}
\alias{run_query}
\alias{run_query}
\alias{run_query_fromfile}
\title{Run a GMQL query}
\usage{
run_query(url, fileName, query, output_gtf = TRUE)

run_query_fromfile(url, fileName, filePath, output_gtf = TRUE)
}
\arguments{
\item{url}{url of server: It must contain the server address 
and base url; service name is added automatically}

\item{fileName}{name of the file}

\item{query}{text of the query}

\item{output_gtf}{logical value indicating file format used for 
storing samples generated by the query.
The possiblities are: 
\itemize{
\item{GTF}
\item{TAB}
}}

\item{filePath}{string path of txt files containing a GMQL query}
}
\value{
None
}
\description{
It runs the GMQL query into repository taken from file or insert as text 
string, using the proper GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{
remote_url = "http://genomic.elet.polimi.it/gmql-rest-r/"
login_gmql(remote_url)
run_query(remote_url, "query_1", "DATASET = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATASET INTO RESULT_DS;", output_gtf = FALSE)

test_path <- system.file("example", package = "RGMQL")
test_query <- file.path(test_path, "query1.txt")
run_query(remote_url, "query_1", test_query, output_gtf = FALSE)
}

}
