% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{run_query_fromfile}
\alias{run_query_fromfile}
\title{Run GMQL query}
\usage{
run_query_fromfile(url, fileName, filePath, output_gtf = TRUE)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{fileName}{string name of the file}

\item{filePath}{string path of txt files containing a GMQL query}

\item{output_gtf}{logical value indicating file format used for 
storing samples generated by the query.
The possiblities are: 
\itemize{
\item{GTF}
\item{TAB}
}}
}
\value{
None

None
}
\description{
It runs a GMQL query from file
}
\details{
If error occures a specific error is printed

If error occures a specific error is printed
}
\examples{

## run query: output GTF

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
test_path <- system.file("example", package = "RGMQL")
test_query <- file.path(test_path, "query1.txt")
run_query(remote_url, "query_1", test_query, output_gtf = FALSE)

}
