% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{runQuery}
\alias{runQuery}
\title{Run GMQL query}
\usage{
runQuery(url, fileName, query, output_gtf = TRUE)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{fileName}{single string name of the file}

\item{query}{single string text of the query}

\item{output_gtf}{logical: file format for samples generated from query
the possiblities are: GTF or TAB}
}
\value{
None
}
\description{
It runs a GMQL query as single string
}
\details{
If error occures a specific error is printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
runQuery(PolimiUrl, "query_1", "DATA_SET_VAR = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATA_SET_VAR INTO RESULT_DS;", output_gtf = FALSE)

}
