% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{sample_region}
\alias{sample_region}
\title{Shows regions data from a dataset sample}
\usage{
sample_region(url, datasetName, sampleName)
}
\arguments{
\item{url}{string url of server: it must contain the server address
and base url; service name is added automatically}

\item{datasetName}{string name of dataset to get}

\item{sampleName}{string sample name to get}
}
\value{
GRanges data containing regions of sample
}
\description{
It retrieves regions data for a specific sample 
(whose name is specified in the paramter "sampleName")
in a specific dataset 
(whose name is specified in the parameter "datasetName") 
using the proper GMQL web service available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

\dontrun{

## Login to GMQL REST services suite as guest

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
login_gmql(remote_url)

## This statement retrieves regions data for sample "S_00000" from public 
## dataset "Example_Dataset1"
 
sample_region(remote_url, "public.Example_Dataset1", "S_00000")

}

}
