% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{save_query}
\alias{save_query}
\alias{save_query_fromfile}
\title{Save GMQL query}
\usage{
save_query(url, queryName, queryTxt)

save_query_fromfile(url, queryName, filePath)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{queryName}{string name of query}

\item{queryTxt}{string text of GMQL query}

\item{filePath}{string local file path of txt file containing a GMQL query}
}
\value{
None
}
\description{
It saves a GMQL query into repository, taken from file or inserted as text 
string, using the proper GMQL web service available on a remote server
}
\details{
If you save a query with the same name of another query already stored 
in repository you will overwrite it; if no error occurs prints, 
"Saved" otherwise it prints the content error
}
\examples{

## Login to GMQL REST services suite as guest

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
login_gmql(remote_url)

## This statement saves query with name "dna_query" 

save_query(remote_url, "dna_query", "DATASET = SELECT() HG19_TCGA_dnaseq; 
MATERIALIZE DATASET INTO RESULT_DS;")


test_path <- system.file("example", package = "RGMQL")
test_query <- file.path(test_path, "query1.txt")
save_query_fromfile(remote_url, "query1", test_query)

}
