% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseOp.R
\name{save_query_fromfile}
\alias{save_query_fromfile}
\title{Save GMQL query from file}
\usage{
save_query_fromfile(url, queryName, filePath)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{queryName}{string name of the GMQL query}

\item{filePath}{string local file path of txt file containing a GMQL query}
}
\value{
None
}
\description{
It saves the GMQL query into repository taken from file
using the proper GMQL web service available on a remote server
}
\details{
if you save a query with the same name of an other query already stored
in repository you will overwrite it,
if no error occures print "Saved" otherwise print the content error
}
\examples{

test_path <- system.file("example", package = "RGMQL")
test_query <- file.path(test_path, "query1.txt")

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
save_query_fromfile(remote_url, "query1", test_query)

}
\seealso{
\code{\link{save_query}}
}
