% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseOp.R
\name{saveQuery.fromfile}
\alias{saveQuery.fromfile}
\title{Save GMQL query from file}
\usage{
saveQuery.fromfile(url, queryName, filePath)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{queryName}{single string name of the GMQL query}

\item{filePath}{single string local file path containing a GMQL query}
}
\value{
None
}
\description{
It saves the GMQL query into repository taken from file
using the proper GMQL web service available on a remote server
}
\details{
if you save a query with the same name of an other query already stored in repository
you will overwrite it,
if no error occures print "Saved" otherwise print the content error
}
\examples{

test_path <- system.file("example",package = "GMQL")
test_query <- file.path(test_path, "query1.txt")

PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl)
saveQuery.fromfile(PolimiUrl,"query1", test_query)

}
\references{
Please, read for writing a GMQL query
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
\seealso{
\code{\link{saveQuery}}
}
