% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseOp.R
\name{saveQuery}
\alias{saveQuery}
\title{Save GMQL query}
\usage{
saveQuery(url, queryName, queryTxt)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{queryName}{single string name of query}

\item{queryTxt}{single string text of GMQL query}
}
\value{
None
}
\description{
It saves the GMQL query into repository
using the proper GMQL web service available on a remote server
}
\details{
if you save a query with the same name of an other query already stored in repository
you will overwrite it,
if no error occures print "Saved" otherwise print the content error
}
\examples{


PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl)
saveQuery(PolimiUrl,"try_1", queryTxt = "DATA_SET_VAR = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATA_SET_VAR INTO RESULT_DS;")

}
\references{
Please, read for writing a GMQL query
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
\seealso{
\code{\link{showQueries}} \code{\link{saveQuery.fromfile}}
}
