% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{show_datasets_list}
\alias{show_datasets_list}
\title{Show datasets}
\usage{
show_datasets_list(url)
}
\arguments{
\item{url}{single string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
List of datasets. Every dataset in the list is described by:
\itemize{
\item{name: name of dataset}
\item{owner: public or name of the user}
}
}
\description{
It shows all GMQL datasets stored by the user or public in remote 
repository, using the proper GMQL web service available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

## Login to GMQL REST services suite as guest

remote_url = "http://www.gmql.eu/gmql-rest/"
login_gmql(remote_url)

## List all datasets

list <- show_datasets_list(remote_url)

}
