% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{show_datasets_list}
\alias{show_datasets_list}
\title{Show Dataset}
\usage{
show_datasets_list(url)
}
\arguments{
\item{url}{single string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
list of datasets.

Every dataset in the list is identified by:
\itemize{
\item{name: name of dataset}
\item{owner: public or name of the user}
}
}
\description{
It show all GMQL dataset stored in repository using the proper GMQL 
web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

## show dataset when logged as guest

remote_url <- "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
list <- show_datasets_list(remote_url)

}
\seealso{
\code{\link{deleteDataset}}
}
