% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{show_job_log}
\alias{show_job_log}
\title{Show a job log}
\usage{
show_job_log(url, job_id)
}
\arguments{
\item{url}{string url of server: It must contain the server address
and base url; service name is added automatically}

\item{job_id}{string id of the job}
}
\value{
log text
}
\description{
It show a job log for specific job
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{

## login with guest user
remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)

## list all jobs
list_jobs <- show_jobs_list(remote_url)
jobs_1 <- list_jobs$jobs[[1]]

## show log
show_job_log(remote_url, jobs_1)

}

}
