% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{show_queries_list}
\alias{show_queries_list}
\title{Show all queries}
\usage{
show_queries_list(url)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
List of queries. Every query in the list is described by:
\itemize{
\item{name: name of query}
\item{text: text of GMQL query}
}
}
\description{
It shows all the GMQL queries saved by the user on remote repository, 
using the proper GMQL web service available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

## Login to GMQL REST services suite
remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
login_gmql(remote_url)

## List all queries executed on remote GMQL system 
list <- show_queries_list(remote_url)

}
