% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{show_samples_list}
\alias{show_samples_list}
\title{Show dataset samples}
\usage{
show_samples_list(url, datasetName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{datasetName}{name of dataset containing the samples whose list we 
like to get; if the dataset is a public dataset, we have to add "public." 
as prefix, as shown in the example below, otherwise no prefix is needed}
}
\value{
List of samples in dataset. Every sample in the list is described 
by:
\itemize{
\item{id: id of sample}
\item{name: name of sample}
\item{path: sample repository path}
}
}
\description{
It show all samples from a specific GMQL dataset on remote repository, 
using the proper GMQL web service available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

## Login to GMQL REST services suite as guest

remote_url = "http://www.gmql.eu/gmql-rest/"
login_gmql(remote_url)

## This statement shows all samples present into public dataset 
## 'Example_Dataset_1'

list <- show_samples_list(remote_url, "public.Example_Dataset_1")

}
