% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{show_schema}
\alias{show_schema}
\title{Show dataset schema}
\usage{
show_schema(url, datasetName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{datasetName}{name of dataset to get
if the dataset is a public dataset, we have to add "public." as prefix, 
as shown in the example below otherwise no prefix is needed}
}
\value{
list of region schema fields.

Every field in the list is identified by:
\itemize{
\item{name: name of field (e.g. chr, start, end, strand ...)}
\item{fieldType: (e.g STRING, DOUBLE ...)}
}
}
\description{
It shows the region attribute schema of a specific GMQL dataset using 
the proper GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

### show schema of public dataset
remote_url <- "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
list <- show_schema(remote_url, "public.GRCh38_ENCODE_BROAD_MAY_2017")

}
