% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{showJobLog}
\alias{showJobLog}
\title{Show a job log}
\usage{
showJobLog(url, job_id)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{job_id}{single string id of the job}
}
\value{
log text
}
\description{
It show a job log for specific job
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{
## login with test user
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl,"test101","test")
## list all jobs
list_jobs <- showJobs(PolimiUrl)
jobs_1 <- list_jobs$jobs[[1]]
## show log
showJobLog(PolimiUrl,jobs_1)
}

}
