% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{showJobs}
\alias{showJobs}
\title{Show all jobs}
\usage{
showJobs(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}
}
\value{
list of jobs
Every job in the list is identified by:
\itemize{
\item{id: id of job used for log, trace and stop}
}
}
\description{
It show all Jobs (runnig, succeded or failed) run by user
}
\details{
If error occures a specific error is printed
}
\examples{

PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl)
list_jobs <- showJobs(PolimiUrl)

}
\seealso{
\code{\link{showJobLog}} @seealso \code{\link{stopJob}} @seealso \code{\link{traceJob}}
}
