% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseOp.R
\name{showQueries}
\alias{showQueries}
\title{Shows all Queries}
\usage{
showQueries(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}
}
\value{
list of queries
Every query in the list is identified by:
\itemize{
\item{name: name of query}
\item{text: text of GMQL query}
}
}
\description{
It shows all the GMQL query saved on repository 
using the proper GMQL web service available on a remote server
}
\details{
if error occures, a specific error is printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
list <- showQueries(PolimiUrl)

}
\seealso{
\code{\link{saveQuery}}
}
