% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{SQRT}
\alias{SQRT}
\title{OPERATOR object class constructor}
\usage{
SQRT(value)
}
\arguments{
\item{value}{string identifying name of region attribute}
}
\value{
SQRT operator object
}
\description{
This class constructor is used to create instances of OPERATOR object,
to be used in GMQL functions that require operator on value.
It prepared input parameter to be passed to library function sqrt,
performing all the type conversion needed
}
\examples{

## This statement allows to build an output dataset out such that all 
## the samples from the input dataset exp are conserved, 
## as well as their region attributes (and their values) 
## and their metadata attributes (and their values). 
## The new metadata attribute concSq is added to all output samples 
## with value correspondent to the mathematical squared root 
## of the pre-existing metadata attribute concentration.

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)
out = subset(exp, metadata_update = list(concSq = SQRT("concentration")))

}
\seealso{
\code{\link{NIL}} \code{\link{SQRT}}
}
