% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-class.R
\name{STD}
\alias{STD}
\title{AGGREGATES object class constructor}
\usage{
STD(value)
}
\arguments{
\item{value}{string identifying name of region attribute}
}
\value{
STD aggregate object
}
\description{
This class constructor is used to create instances of AGGREGATES object,
to be used in GMQL functions that require aggregate on value.
It prepares input parameter to be passed to the library function 
standard deviation, performing all the type conversions needed
}
\examples{

### local with CustomParser
init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)

## This statement copies all samples of exp into res dataset, 
## and then calculates new metadata attributes for each of them: 
## std_score is the standard deviation score of the sample regions.

res = extend(input_data = exp, list(std_score = STD("score")))

}
\seealso{
\code{\link{SUM}} \code{\link{COUNT}} \code{\link{MIN}} 
\code{\link{MAX}} \code{\link{AVG}} \code{\link{MEDIAN}} 
\code{\link{BAG}} \code{\link{BAGD}}
\code{\link{Q1}} \code{\link{Q2}} \code{\link{Q3}}
}
