% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_init.R
\name{stop_gmql}
\alias{stop_gmql}
\title{Stop GMQL server}
\usage{
stop_gmql()
}
\value{
None
}
\description{
it stops GMQL server processing
}
\examples{

## These statements first initializes GMQL with local processing and with 
## sample files output format as tab-delimited, and then stops it

init_gmql("tab", FALSE)

stop_gmql()

}
