% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{stop_job}
\alias{stop_job}
\title{Stop a job}
\usage{
stop_job(url, job_id)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{job_id}{string id of the job}
}
\value{
None
}
\description{
It stops a specific current job
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
list_jobs <- show_jobs_list(remote_url)
jobs_1 <- list_jobs$jobs[[1]]
stop_job(remote_url, jobs_1)
}

}
