% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{stopJob}
\alias{stopJob}
\title{Stop a job}
\usage{
stopJob(url, job_id)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{job_id}{single string id of the job}
}
\value{
None
}
\description{
It stops a specific current job
}
\details{
If error occures a specific error is printed
}
\examples{

\dontrun{
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl,"test101","test")
list_jobs <- showJobs(PolimiUrl)
jobs_1 <- list_jobs$jobs[[1]]
stopJob(PolimiUrl,jobs_1)
}

}
