% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gmql_materialize.R
\docType{methods}
\name{take}
\alias{take}
\alias{take,GMQLDataset-method}
\alias{take-method}
\title{Method take}
\usage{
take(data, ...)

\S4method{take}{GMQLDataset}(data, rows = 0L)
}
\arguments{
\item{data}{returned object from any GMQL function}

\item{...}{Additional arguments for use in specific methods}

\item{rows}{number of rows for each sample regions that you want to 
retrieve and stored in memory.
by default is 0 that means take all rows for each sample}
}
\value{
GrangesList with associated metadata
}
\description{
GMQL Operation: TAKE

It saves the contents of a dataset that contains samples metadata 
and samples regions.
It is normally used to store in memoery the contents of any dataset 
generated during a GMQL query. the operation can be very time-consuming.
If you have invoked any materialization before take function, 
all those dataset will be materialized as folder.
}
\examples{

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
rd = read_dataset(test_path)
aggr = aggregate(rd, DF("antibody_target", "cell_karyotype"))
taken <- take(aggr, rows = 45)

}
