% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQL4TFarm.R
\name{TFARMemAtrix}
\alias{TFARMemAtrix}
\title{GRangesList to TFARM matrix}
\usage{
TFARMemAtrix(rangesList, metadata = NULL, metadata_prefix = NULL,
  regions = NULL)
}
\arguments{
\item{rangesList}{GrangesList object}

\item{metadata}{vector of character containing names of metadata to be searched for in metdata files.
data will be extracted if at least one condition is satisfied:
this condition will be logically "ANDed" with prefix filtering (see below )
if NULL no filtering action occured (i.e every sample will be taken for regions filtering)}

\item{metadata_prefix}{vector of character that will filter metadata containing rispectively every element
of this vector.
number of elelment in both vector must match}

\item{regions}{vector of character that will extract only region attribute specified
if NULL no regions will be taken and the output will be only GRanges made up by the first attribute
(seqnames,start,end,strand)}
}
\value{
granges with selected regions (if any) in elementMetadata
}
\description{
Auxiliary function to interface GMQL suite to an exeternal package called TFARM
It prepares data in the correct format for TFARM processing
}
\examples{


test_path <- system.file("example","DATA_SET_VAR_GTF",package = "RGMQL")
grl <- importGMQL.gtf(test_path)
TFARMemAtrix(grl,regions = c("pvalue","peak"))


}
