% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{trace_job}
\alias{trace_job}
\title{Trace a job}
\usage{
trace_job(url, job_id)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name will be added automatically}

\item{job_id}{string id of the job}
}
\value{
text trace log
}
\description{
It traces a specific current job
}
\details{
If error occures a specific error is printed
}
\examples{
remote_url = "http://130.186.13.219/gmql-rest"
\dontrun{
login_gmql(remote_url)
list_jobs <- show_jobs_list(remote_url)
jobs_1 <- list_jobs$jobs[[1]]
trace_job(remote_url, jobs_1)
}


}
