% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{traceJob}
\alias{traceJob}
\title{Trace a job}
\usage{
traceJob(url, job_id)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{job_id}{single string id of the job}
}
\value{
text trace log
}
\description{
It traces a specific current job
}
\details{
If error occures a specific error is printed
}
\examples{
\dontrun{
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl,"test101","test")
list_jobs <- showJobs(PolimiUrl)
jobs_1 <- list_jobs$jobs[[1]]
traceJob(PolimiUrl,jobs_1)
}


}
