% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_union.R
\docType{methods}
\name{union}
\alias{union}
\alias{union,GMQLDataset,GMQLDataset-method}
\title{Method union}
\usage{
\S4method{union}{GMQLDataset,GMQLDataset}(x, y)
}
\arguments{
\item{x}{GMQLDataset object}

\item{y}{GMQLDataset object}
}
\value{
GMQLDataset object. It contains the value to use as input 
for the subsequent GMQLDataset method
}
\description{
Wrapper to GMQL UNION operator

It is used to integrate homogeneous or heterogeneous samples 
of two datasets within a single dataset; for each sample of either input 
dataset, a result sample is created as follows:
\itemize{
\item {Metadata are the same as in the original sample.}
\item {Resulting schema is is the schema of the left input dataset. }
\item {Regions are the same (in coordinates and attribute values) 
as in the original sample if it is from the left input dataset; 
if it is from the right input dataset, its regions are the same in 
coordinates, but only region attributes identical (in name and type) to 
those of the first input dataset are retained, with the same values.
Region attributes which are missing in an input dataset sample 
w.r.t. the merged schema are set to null.}
}
}
\examples{

## Thi statement initializes and runs the GMQL server for local execution 
## and creation of results on disk. Then, with system.file() it defines 
## the path to the folders "DATASET" and "DATASET_GDM" in the subdirectory 
## "example" of the package "RGMQL" and opens such folder as a GMQL 
## dataset named "data1" and "data2" respectively using customParser

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
test_path2 <- system.file("example", "DATASET_GDM", package = "RGMQL")
data1 <- read_dataset(test_path)
data2 <- read_dataset(test_path2)

## This statement creates a dataset called 'full' which contains all samples 
## from the datasets 'data1' and 'data2'

res <- union(data1, data2)


}
