% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{UP}
\alias{UP}
\title{DISTAL object class constructor}
\usage{
UP()
}
\value{
None
}
\description{
This class constructor is used to create instances of DISTAL object
to be used in GMQL functions \code{\link{join}} in genometric predicate parameter
that require distal condition on value
In this case UP: denotes the upstream direction of the genome.
They are interpreted as predicates that must hold on the regions of the experiment; 
UP is true when region of experiment is in the upstream genome of the anchor region.
When this clause is not present, distal conditions apply to both the directions of the genome.
}
\examples{

""

}
\seealso{
\code{\link{DLE}} \code{\link{DGE}}  \code{\link{DOWN}} \code{\link{MD}}
}
