% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{uploadSamples}
\alias{uploadSamples}
\title{Upload dataset}
\usage{
uploadSamples(url, datasetName, folderPath, schemaName = NULL,
  isGMQL = TRUE)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{datasetName}{name of dataset to get}

\item{folderPath}{single string local path to the folder containing the samples}

\item{schemaName}{single string name of schema used to parse the samples
schemaName available are:
\itemize{
\item{NARROWPEAK}
\item{BROADPEAK}
\item{VCF}
\item{BED}
\item{BEDGRAPH}
}
if schema is NULL it's looking for a XML schema file to read}

\item{isGMQL}{single logical value indicating whether is GMQL dataset or not}
}
\value{
None
}
\description{
It uploads a folder (GMQL or not) containing sample files using the proper GMQL 
web service available on a remote server: a new dataset is created on repository
}
\details{
If no error occures print "Upload Complete", otherwise a specific error is printed
}
\examples{

\dontrun{

### upload of GMQL dataset with no schema selection
test_path <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(url = PolimiUrl)
uploadSamples(PolimiUrl,"dataset1",folderPath = test_path)
}

}
