% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{ASC}
\alias{ASC}
\title{ORDER object class constructor}
\usage{
ASC(value)
}
\arguments{
\item{value}{single string identifying name of metadata and/or region attribute}
}
\value{
ASC ordering object
}
\description{
This class constructor is used to create instances of ORDER object
to be used in GMQL functions that require ordering on value.
It define an ascending order for input value
}
\examples{

## it get the first 5 samples on the basis of their region counter, 
## those with the smaller RegionCount and then for each of them, 7 regions on the basis of 
## their score, those with the higher score

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "RGMQL")
r = readDataset(test_path)
o = order(r,list(ASC("Region_Count")), mtop = 5,regions_ordering = list(DESC("score")),rtop=7)

}
