% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{ASC}
\alias{ASC}
\title{ORDER object class constructor}
\usage{
ASC(value)
}
\arguments{
\item{value}{string identifying name of metadata or region attribute}
}
\value{
ASC ordering object
}
\description{
This class constructor is used to create instances of ORDER object,
to be used in GMQL functions that require ordering on value.
It defines a descending order for input value
}
\examples{

## It gets the first 5 samples on the basis of their region count, 
## those with the smaller Region_Count and then for each of them, 
## It gets 7 regions on the basis of their score, those with the 
## higher score

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
r = read_dataset(test_path)
o = sort(r,TRUE, list(ASC("Region_Count")),list(DESC("score")),
fetch_opt = "mtop", num_fetch = 5, reg_fetch_opt = "rtop", 
reg_num_fetch = 7)

}
