% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-class.R
\name{AVG}
\alias{AVG}
\title{AGGREGATES object class constructor}
\usage{
AVG(value)
}
\arguments{
\item{value}{string identifying name of metadata or region attribute}
}
\value{
AVG aggregate object
}
\description{
This class constructor is used to create instances of AGGREGATES object,
to be used in GMQL functions that require aggregate on value.
It prepares input parameter to be passed to the library function aritmetic
mean, performing all the type conversions needed
}
\examples{

### local with CustomParser
init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)

## The following cover operation produces output regions where at least 2 
## and at most 3 regions ofexp overlap, having as resulting region 
## attributes the average signal of the overlapping regions; 
## the result has one sample for each input cell.

res = cover(exp, 2, 3, c("cell"), 
list(avg_signal = AVG("signal")))

}
\seealso{
\code{\link{SUM}} \code{\link{COUNT}} \code{\link{MIN}} 
\code{\link{MAX}} \code{\link{MEDIAN}} \code{\link{STD}} 
\code{\link{BAG}} \code{\link{BAGD}}
\code{\link{Q1}} \code{\link{Q2}} \code{\link{Q3}}
}
