% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{compile_query}
\alias{compile_query}
\title{Compile GMQL query}
\usage{
compile_query(url, query)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{query}{string text of the query}
}
\value{
None
}
\description{
It compiles a GMQL query as single string using the proper GMQL web service 
available on a remote server
}
\examples{

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
compile_query(remote_url, "DATASET = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATASET INTO RESULT_DS;")

}
