% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-functions.R
\name{Evaluation-Function}
\alias{Evaluation-Function}
\alias{FN}
\alias{Evaluation-Function}
\alias{EX}
\alias{Evaluation-Function}
\alias{DF}
\title{Condition evaluation functions}
\usage{
FN(...)

EX(...)

DF(...)
}
\arguments{
\item{...}{series of string identifying name of metadata attribute 
to be evaluated}
}
\value{
list of 2-D array containing method of evaluation and metadata
}
\description{
These functions are used to support joinBy and/or groupBy function parameter.
It create a list of one element: matrix containing the two coloumn:
type of condition evaluation and the metadata attribute name
}
\details{
\itemize{
\item{FN: It defines a FULL (FULLNAME) evaluation of the input values.
FULL evaluation: two attributes match if they both end with value and,
if they have further prefixes, the two prefix sequences are identical}
\item{EX: It defines a EXACT evaluation of the input values.
EXACT evaluation: only attributes exactly as value will match; 
no further prefixes are allowed. }
\item{DF: It defines a DEFAULT evaluation of the input values.
DEFAULT evaluation: the two attributes match if both end with value.}
}
}
\examples{

"where is my example?"

}
