% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-class.R
\name{COUNT}
\alias{COUNT}
\title{AGGREGATES object class constructor}
\usage{
COUNT()
}
\value{
COUNT aggregate object
}
\description{
This class constructor is used to create instances of AGGREGATES object,
to be used in GMQL functions that require aggregate on value.
It prepares input parameter to be passed to the library function count,
performing all the type conversions needed
}
\examples{

## local with CustomParser
init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)

## counts the regions in each sample and stores their number as value 
## of the new metadata RegionCount attribute of the sample.

out = extend(input_data = exp, list(RegionCount = COUNT()))

}
\seealso{
\code{\link{SUM}} \code{\link{MIN}} 
\code{\link{MAX}} \code{\link{AVG}} \code{\link{MEDIAN}} 
\code{\link{STD}} \code{\link{BAG}} \code{\link{BAGD}}
\code{\link{Q1}} \code{\link{Q2}} \code{\link{Q3}}
}
