% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete dataset}
\usage{
delete_dataset(url, datasetName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{datasetName}{string name of dataset to delete}
}
\value{
None
}
\description{
It deletes single private dataset specified by name from repository 
using the proper GMQL web service available on a remote server
}
\details{
If no error occur, print "Deleted Dataset", otherwise a specific error 
is printed
}
\examples{

\dontrun{

## This dataset does not exist

remote_url <- "http://genomic.elet.polimi.it/gmql-rest-r/"
login_gmql(remote_url)
delete_dataset(remote_url, "job_test1_test101_20170604_180908_RESULT_DS")

}

}
