% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{DESC}
\alias{DESC}
\title{ORDER object class constructor}
\usage{
DESC(value)
}
\arguments{
\item{value}{string identifying name of metadata or region attribute}
}
\value{
DESC ordering object
}
\description{
This class constructor is used to create instances of ORDER object,
to be used in GMQL functions that require ordering on value.
It defines a descending order for input value
}
\examples{

## It orders the samples according to the Region_count metadata attribute 
## and takes the two samples that have the highest count. 

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
r = read_dataset(test_path)
o = sort(r,TRUE, c("Region_Count"), fetch_opt = "mtop", 
num_fetch = 2)
 
}
