% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{DESC}
\alias{DESC}
\title{ORDER object class constructor}
\usage{
DESC(value)
}
\arguments{
\item{value}{single string identifying name of metadata and/or region attribute}
}
\value{
DESC ordering object
}
\description{
This class constructor is used to create instances of ORDER object
to be used in GMQL functions that require ordering on value.
It define a descending order for input value
}
\examples{

## it orders the samples according to the Region_count metadata attribute and takes the two samples 
## that have the highest count. 

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = readDataset(test_path)
o = order(r,list(DESC("Region_Count")), mtop = 2)

}
