% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DGE}
\alias{DGE}
\title{DISTAL object class constructor}
\usage{
DGE(value)
}
\arguments{
\item{value}{single string identifying distance between genomic regions in base pairs,}
}
\value{
None
}
\description{
This class constructor is used to create instances of DISTAL object
to be used in GMQL functions \code{\link{join}} in genometric predicate parameter
that require distal condition on value
In this case DGE: denotes the greater-equal distance clause, which selects all the regions of the experiment such
that their distance from the anchor region is greater than, or equal to, 'value' bases.
}
\examples{
\dontrun{

}
""
}
\seealso{
\code{\link{DLE}} \code{\link{MD}}  \code{\link{DOWN}} \code{\link{UP}}
}
