% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DLE}
\alias{DLE}
\title{DISTAL object class constructor}
\usage{
DLE(value)
}
\arguments{
\item{value}{single string identifying distance between genomic regions in base pairs,}
}
\value{
None
}
\examples{

HM_TSS = JOIN(MD(1), DLE(120000); output: RIGHT; joinby: provider) TSS HM;
### Given a dataset 'hm' and one called 'tss' with a sample including Transcription Start Site annotations,
it searches for those regions of hm that are at a minimal distance from a transcription start site (TSS) 
and takes the first/closest one for each TSS, 
provided that such distance is lesser than 120K bases and joined 'tss' and 'hm' samples are obtained 
from the same provider (joinby clause).

#' initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
test_path2 <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
TSS = readDataset(test_path)
HM = readDataset(test_path2)
join_data = join(tss,hm,genometric_predicate=list(list(MD("1"),DLE("120000"))),c("provider"),region_output="RIGHT")

}
\seealso{
\code{\link{DGE}} \code{\link{MD}}  \code{\link{DOWN}} \code{\link{UP}}
}
