% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DOWN}
\alias{DOWN}
\title{DISTAL object class constructor}
\usage{
DOWN()
}
\value{
None
}
\description{
This class constructor is used to create instances of DISTAL object
to be used in GMQL functions \code{\link{join}} in genometric predicate parameter
that require distal condition on value
In this case DOWN: denotes the downstream direction of the genome.
They are interpreted as predicates that must hold on the regions of the experiment; 
DOWN is true when region of experiment is in the downstream genome of the anchor region.
When this clause is not present, distal conditions apply to both the directions of the genome.
}
\examples{

HM_TSS = JOIN(MD(1), DLE(120000); output: RIGHT; joinby: provider) TSS HM;
### Given a dataset 'hm' and one called 'tss' with a sample including Transcription Start Site annotations,
it searches for those regions of hm that are at a minimal distance from a transcription start site (TSS) 
and takes the first/closest one for each TSS, 
provided that such distance is greater than 12K bases and joined 'tss' and 'hm' samples are obtained 
from the same provider (joinby clause).

#' initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
test_path2 <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
TSS = readDataset(test_path)
HM = readDataset(test_path2)
join_data = join(tss,hm,genometric_predicate=list(list(MD("1"),DGE("12000"),DOWN)),c("provider"),region_output="RIGHT")


}
\seealso{
\code{\link{DLE}} \code{\link{DGE}}  \code{\link{UP}} \code{\link{MD}}
}
