% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{download_dataset}
\alias{download_dataset}
\title{Download Dataset}
\usage{
download_dataset(url, datasetName, path = getwd())
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name will be added automatically}

\item{datasetName}{string name of dataset we want to get}

\item{path}{string local path folder where store dataset,
by defualt is R working directory}
}
\value{
None
}
\description{
It donwloads private dataset as zip file from repository to local path 
specified using the proper GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

#### download dataset in r working directory
#### in this case we try to download public dataset

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
download_dataset(remote_url, "public.HG19_BED_ANNOTATION", path = getwd())

}
