% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{download_dataset}
\alias{download_dataset}
\alias{download_as_GRangesList}
\title{Download Dataset}
\usage{
download_dataset(url, datasetName, path = getwd())

download_as_GRangesList(url, datasetName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name are added automatically}

\item{datasetName}{string name of dataset we want to get}

\item{path}{string local path folder where to store dataset,
by default it is R working directory}
}
\value{
None

GRangesList containing all GMQL sample in dataset
}
\description{
It donwloads private dataset as zip file from remote repository to local 
path, or donwloads and saves it into R environment as GRangesList using 
the proper GMQL web service available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

## Download dataset in r working directory
## in this case we try to download public dataset

\dontrun{

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r"
login_gmql(remote_url)
download_dataset(remote_url, "public.Example_Dataset1", path = getwd())

## Create GRangesList from public dataset Example_Dataset1 got 
## from repository

download_as_GRangesList(remote_url, "public.Example_Dataset1")
}

}
