% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download Dataset}
\usage{
downloadDataset(url, datasetName, path = getwd())
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{single string name of dataset we want to get}

\item{path}{single string local path folder where store dataset,
by defualt is R working directory}
}
\value{
None
}
\description{
It donwloads private dataset as zip file from repository to local path specified 
using the proper GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

#### download dataset in r working directory
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl,"test101","test")
downloadDataset(PolimiUrl,"dataset_test",path = getwd())

}
