% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{downloadDatasetToGrangesList}
\alias{downloadDatasetToGrangesList}
\title{Download Dataset in GrangesList}
\usage{
downloadDatasetToGrangesList(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}
}
\value{
GrangesList containing all GMQL sample in dataset
}
\description{
It donwloads private dataset from repository to GrangesList
}
\details{
If error occured a specific error will be printed
}
\examples{

#### create grangeslist from dataset in repository
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl,"test101","test")
downloadDatasetToGrangesList(PolimiUrl,"dataset_test")

}
