% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Materialize.R
\name{execute}
\alias{execute}
\title{GMQL Function: EXECUTE}
\usage{
execute()
}
\value{
None
}
\description{
Execute GMQL query.
The function works only after invoking at least one materialize
}
\examples{

init_gmql()
test_path <- system.file("example","DATASET",package = "RGMQL")
r = read_dataset(test_path)
s = filter(r)
m = aggregate(s, groupBy = c("antibody_targer","cell_karyotype"))
materialize(m, dir_out = test_path)

\dontrun{
execute()
}
}
