% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_materialize.R
\name{execute}
\alias{execute}
\title{GMQL Function: EXECUTE}
\usage{
execute()
}
\value{
None
}
\description{
It executes GMQL query.
The function works only after invoking at least one collect
}
\examples{
## This statement initializes and runs the GMQL server for local execution 
## and creation of results on disk. Then, with system.file() it defines 
## the path to the folder "DATASET" in the subdirectory "example" 
## of the package "RGMQL" and opens such folder as a GMQL dataset 
## named "data"

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
data = read_gmql(test_path)

## The following statement materializes the dataset "data", previoulsy read, 
## at the specific destination test_path into local folder "ds1" opportunely 
## created

collect(data, dir_out = test_path)

## This statement executes GMQL query.
\dontrun{

execute()
}
}
