% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Materialize.R
\name{execute}
\alias{execute}
\title{GMQL Function: EXECUTE}
\usage{
execute()
}
\value{
None
}
\description{
execute GMQL query.
The function works only after invoking at least one materialize
}
\details{
After invoking execution function, all varialbe associated to DAG will be removed
from scala enviroment, although the associated R variable will remain stored in R environment
}
\examples{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = readDataset(test_path)
s = select(input_data = r)
m = merge(groupBy = c("antibody_targer","cell_karyotype"),input_data = s)
materialize(input_data = m, dir_out = test_path)
execute()

}
