% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-class.R
\name{FULL}
\alias{FULL}
\title{CONDITION object class constructor}
\usage{
FULL(value)
}
\arguments{
\item{value}{string identifying name of metadata and/or 
region attribute to be evaluated}
}
\value{
FULL condition object
}
\description{
This class constructor is used to create instances of CONDITION object
to be used in GMQL functions that require evaluation on value.
This constructor defines a FULL (FULLNAME) evaluation of the input value.
FULL evaluation: two attributes match if they both end with value and,
if they have further prefixes, the two prefix sequences are identical
}
\examples{

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
r = read_dataset(test_path)

## select with condition
## the first and the third attribute are default the second one is FULL
s = select(r, semi_join = list("cell_type", FULL("cell"), "attribute_tag"),
semi_join_dataset = r)

\dontrun{

## select with default condition
## Tthe FULL condition is treated as default due to coercion
s = select(r, semi_join = c("cell_type", "cell", FULL("attribute_tag")),
semi_join_dataset = r)

}

}
\seealso{
\code{\link{EXACT}}
}
