% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLDataset-class.R
\docType{class}
\name{GMQLDataset-class}
\alias{GMQLDataset-class}
\alias{GMQLDataset}
\title{Class GMQLDataset}
\usage{
GMQLDataset(value)
}
\arguments{
\item{value}{value associated to GMQL dataset}
}
\value{
instance of GMQL dataset
}
\description{
Abstract class representing GMQL dataset
}
\section{Slots}{

\describe{
\item{\code{value}}{value associated to GMQL dataset}
}}

