% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_group.R
\docType{methods}
\name{group_by}
\alias{group_by}
\alias{group_by,GMQLDataset-method}
\title{Method group_by}
\usage{
\S4method{group_by}{GMQLDataset}(.data, groupBy_meta = NULL,
  groupBy_regions = NULL, region_aggregates = NULL,
  meta_aggregates = NULL)
}
\arguments{
\item{.data}{GMQLDataset object}

\item{groupBy_meta}{it define condition evaluation on metadata.
\itemize{
\item{\code{\link{FN}}: Fullname evaluation, two attributes match 
if they both end with value and, if they have a further prefixes,
the two prefix sequence are identical}
\item{\code{\link{EX}}: Exact evaluation, only attributes exactly 
as value will match; no further prefixes are allowed. }
\item{\code{\link{DF}}: Default evaluation, the two attributes match 
if both end with value.}
}}

\item{groupBy_regions}{vector of string made up by schema field attribute}

\item{region_aggregates}{It accept a series of aggregate function on 
region attribute. 
All the element in the form \emph{key} = \emph{aggregate}.
The \emph{aggregate} is an object of class AGGREGATES
The aggregate functions available are: \code{\link{SUM}}, 
\code{\link{COUNT}}, \code{\link{MIN}}, \code{\link{MAX}}, 
\code{\link{AVG}}, \code{\link{MEDIAN}}, \code{\link{STD}}, 
\code{\link{BAG}}, \code{\link{BAGD}}, \code{\link{Q1}}, 
\code{\link{Q2}}, \code{\link{Q3}}.
Every aggregate accepts a string value, execet for COUNT, which does not 
have any value.
Argument of 'aggregate function' must exist in schema, i.e. among region 
attributes. Two style are allowed:
\itemize{
\item list of key-value pairs: e.g. sum = SUM("pvalue")
\item list of values: e.g. SUM("pvalue")
}
"mixed style" is not allowed}

\item{meta_aggregates}{It accept a series of aggregate function on 
metadata attribute.
All the element in the form \emph{key} = \emph{aggregate}.
The \emph{aggregate} is an object of class AGGREGATES
The aggregate functions available are: \code{\link{SUM}}, 
\code{\link{COUNT}}, \code{\link{MIN}}, \code{\link{MAX}}, 
\code{\link{AVG}}, \code{\link{MEDIAN}}, \code{\link{STD}}, 
\code{\link{BAG}}, \code{\link{BAGD}}, \code{\link{Q1}}, 
\code{\link{Q2}}, \code{\link{Q3}}.
Every aggregate accepts a string value, execet for COUNT, which does not 
have any value.
Argument of 'aggregate function' must exist in schema, i.e. among region 
attributes. Two style are allowed:
\itemize{
\item list of key-value pairs: e.g. sum = SUM("cell")
\item list of values: e.g. SUM("cell")
}
"mixed style" is not allowed}
}
\value{
GMQLDataset object. It contains the value to use as input 
for the subsequent GMQLDataset method
}
\description{
Wrapper to GMQL GROUP operator
}
\examples{

}
