% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{import_gmql}
\alias{import_gmql}
\title{Create GRangesList from GMQL Dataset}
\usage{
import_gmql(dataset_path, is_gtf)
}
\arguments{
\item{dataset_path}{string GMQL dataset folder path}

\item{is_gtf}{logical value indicating if samples inside are in GTF format
if TRUE and dataset does not contain gtf sample an error occures}
}
\value{
GrangesList containing all GMQL samples in dataset
}
\description{
It create a GrangesList from GMQL samples in dataset
It reads sample files in GTF or GDM/tabulated format
}
\examples{

test_path <- system.file("example", "DATASET", package = "RGMQL")
grl = import_gmql(test_path, TRUE)


}
\seealso{
\code{\link{export_gmql}}
}
