% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_init.R
\name{init_gmql}
\alias{init_gmql}
\title{Init GMQL server}
\usage{
init_gmql(output_format = "GTF", remote_processing = FALSE, url = NULL,
  username = NULL, password = NULL)
}
\arguments{
\item{output_format}{string identifies the output format of sample files.
It can be TAB, GTF or COLLECT:
\itemize{
\item{TAB: tab delimited file format}
\item{GTF: tab-delimited text fstandard ormat based on the general 
feature format}
\item{COLLECT: used for storing output in memory}
}}

\item{remote_processing}{logical value specifying the processing mode.
True for processing on cluster (remote), false for local processing.}

\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically.
If NULL, no login is performed.
You can always perform it by calling the function \code{\link{login_gmql}} 
explicitly}

\item{username}{string name used during signup}

\item{password}{string password used during signup}
}
\value{
None
}
\description{
It initializes and runs GMQL server for executing GMQL query
It also performs a login to GMQL REST services suite, if needed
}
\examples{

## This statement initializes GMQL with local processing with sample files 
## output format as tab delimited

init_gmql("tab", FALSE)

## initializes GMQL with remote processing

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
init_gmql(remote_processing = TRUE, url = remote_url)

}
