% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_init.R
\name{init_gmql}
\alias{init_gmql}
\title{Init GMQL server}
\usage{
init_gmql(output_format = "GTF", remote_processing = FALSE,
  url = NULL, username = NULL, password = NULL)
}
\arguments{
\item{output_format}{string that identifies the output format of all sample 
files. It can be TAB, GTF or COLLECT:
\itemize{
\item{TAB: tab-delimited file format}
\item{GTF: tab-delimited text standard format based on the General 
Feature Format}
\item{COLLECT: used for storing output in memory (only in the case of local 
processing, i.e., remote_processing = FALSE)}
}}

\item{remote_processing}{logical value specifying the processing mode.
True for processing on cluster (remote), false for local processing.}

\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically.
If NULL, no login is performed.
You can always perform it by calling the function \code{\link{login_gmql}} 
explicitly}

\item{username}{string name used during remote server signup}

\item{password}{string password used during remote server signup}
}
\value{
None
}
\description{
It initializes and runs GMQL server for executing GMQL query.
It also performs a login to GMQL REST services suite, if needed
}
\examples{

## This statement initializes GMQL with local processing with sample files 
## output format as tab-delimited

init_gmql("tab", FALSE)

## This statement initializes GMQL with remote processing

remote_url = "http://www.gmql.eu/gmql-rest/"
init_gmql(remote_processing = TRUE, url = remote_url)

}
