% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{initGMQL}
\alias{initGMQL}
\title{Init GMQL Server}
\usage{
initGMQL(output_format, remote_processing = FALSE)
}
\arguments{
\item{output_format}{single string identifies the output format of sample files.
Can be TAB, GTF or COLLECT
\itemize{
\item{TAB: tab delimited file format}
\item{GTF: file format used to hold information about gene structure.
It is a tab-delimited text format based on the general feature format}
\item{COLLECT: used for storing output in memory}
}}

\item{remote_processing}{logical value specifying the processing mode.
can be local (local machine) or remote (cluster).}
}
\value{
no object return
}
\description{
Initialize and run GMQL server for executing GMQL query
}
\examples{

# initialize GMQL with local processing
library(rscala)
initGMQL("tab",FALSE)

}
