% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{log_job}
\alias{log_job}
\alias{show_job_log}
\alias{trace_job}
\title{Show a job log or trace}
\usage{
show_job_log(url, job_id)

trace_job(url, job_id)
}
\arguments{
\item{url}{string url of server: It must contain the server address
and base url; service name is added automatically}

\item{job_id}{string id of the job}
}
\value{
Log or trace text
}
\description{
It shows a job log or traces a specific job
}
\details{
If error occurs, a specific error is printed
}
\examples{
## Login to GMQL REST services suite as guest

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
login_gmql(remote_url)

## List all jobs
list_jobs <- show_jobs_list(remote_url)

\dontrun{
jobs_1 <- list_jobs$jobs[[1]]

## Show jobs_1 log
show_job_log(remote_url, jobs_1)

## Trace jobs_1
trace_job(remote_url, jobs_1)

}

}
