% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOp.R
\name{login.GMQL}
\alias{login.GMQL}
\title{Login to GMQL}
\usage{
login.GMQL(url, username = NULL, password = NULL)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}

\item{username}{single string name used during signup}

\item{password}{single string password used during signup}
}
\value{
None
}
\description{
Login to GMQL REST services suite as a registered user, specifying username and password,
or as guest using the proper GMQL web service available on a remote server
}
\details{
if both username and password are NULL you will be logged as guest
After login you will receive an authentication token.
As token remains vaild on server (until the next login / registration) a user can safely use a token for
a previous session as a convenience, this token is saved in Global environment to perform
subsequent REST call even on complete R restart (if is environemnt has been saved, of course ...)
If error occures a specific error is printed
}
\examples{

### login as guest
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl)

\dontrun{

### login with username and password
PolimiUrl = "http://130.186.13.219/gmql-rest"
login.GMQL(PolimiUrl,"test101","test")

}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
