% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{login_gmql}
\alias{login_gmql}
\title{Login to GMQL}
\usage{
login_gmql(url, username = NULL, password = NULL)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{username}{string name used during signup}

\item{password}{string password used during signup}
}
\value{
None
}
\description{
Login to GMQL REST services suite as a registered user, specifying username 
and password, or as guest, using the proper GMQL web service available 
on a remote server
}
\details{
If both username and password are NULL you will be logged as guest.
After login you will receive an authentication token.
As token remains vaild on server (until the next login / registration) 
a user can safely use a token for a previous session as a convenience;
this token is saved in R Global environment to perform subsequent REST call 
even on complete R restart (if the environment has been saved).
If error occurs a specific error is printed
}
\examples{
## Login to GMQL REST services suite as guest

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
login_gmql(remote_url)

}
