% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Materialize.R
\name{materialize}
\alias{materialize}
\title{GMQL Operation: MATERIALIZE}
\usage{
materialize(input_data, dir_out = getwd())
}
\arguments{
\item{input_data}{returned object from any GMQL function}

\item{dir_out}{destination folder path.
by default is current working directory of the R process}
}
\value{
None
}
\description{
It saves the contents of a dataset that contains samples metadata and samples regions.
It is normally used to persist the contents of any dataset generated during a GMQL query.
Any dataset can be materialized, but the operation can be very time-consuming.
For best performance, materialize the relevant data only.
}
\examples{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = readDataset(test_path)
s = select(input_data = r)
m = merge(groupBy = c("antibody_targer","cell_karyotype"),input_data = s)
materialize(input_data = m, dir_out = test_path)

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
