% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-class.R
\name{MAX}
\alias{MAX}
\title{AGGREGATES object class constructor}
\usage{
MAX(value)
}
\arguments{
\item{value}{string identifying name of region attribute}
}
\value{
MAX aggregate object
}
\description{
This class constructor is used to create instances of AGGREGATES object,
to be used in GMQL functions that require aggregate on value.
It prepares input parameter to be passed to the library function maximum,
performing all the type conversions needed
}
\examples{

### local with CustomParser
init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)

## This statement copies all samples of exp into res dataset, 
## and then calculates new metadata attributes for each of them: 
## max_score is the maximum score of the sample regions.

res = extend(exp, max_score = MAX("score"))


}
\seealso{
\code{\link{SUM}} \code{\link{COUNT}} \code{\link{MIN}} 
\code{\link{AVG}} \code{\link{MEDIAN}} \code{\link{STD}} 
\code{\link{BAG}} \code{\link{BAGD}}
\code{\link{Q1}} \code{\link{Q2}} \code{\link{Q3}}
}
