% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{META}
\alias{META}
\title{OPERATOR object class constructor}
\usage{
META(value, type = NULL)
}
\arguments{
\item{value}{string identifying name of metadata attribute}
}
\value{
META operator object
}
\description{
This class constructor is used to create instances of OPERATOR object,
to be used in GMQL functions that require operator on value.
It prepared input parameter to be passed to library function meta,
performing all the type conversion needed
}
\examples{

## allows to select, in all input sample, all those regions for which the 
## region attribute score has a value which is greater than the metadata 
## attribute value "avg_score" in their sample.

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)
data = select(exp, region_predicate = score > META("avg_score"));


## It define a new region attribute with the value of a metadata attribute 
## using the syntax region_attribute AS META(metadata_attribute, type)

out = project(exp, regions_update = list(signal = META(avg_signal, DOUBLE)))


}
\seealso{
\code{\link{NIL}} \code{\link{SQRT}}
}
