% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{order}
\alias{order}
\title{GMQL operation: ORDER}
\usage{
order(input_data, metadata_ordering = NULL, mtop = 0, mtopg = 0,
  mtopp = 0, regions_ordering = NULL, rtop = 0, rtopg = 0, rtopp = 0)
}
\arguments{
\item{input_data}{returned object from any GMQL function}

\item{metadata_ordering}{list of ORDER objects where every object 
contains the name of metadata.
The ORDER's available are: \code{\link{ASC}}, \code{\link{DESC}}
Every condition accepts only one string value. (e.g. ASC("cell_type") )}

\item{mtop}{integer value specifying the first k samples to retrieve.
default is 0 that means every sample must be considered}

\item{mtopg}{integer value specifying the first k samples to retrieve 
in each group. 
default is 0 that means every sample must be considered}

\item{mtopp}{integer value specifying the percentage of samples to retrieve.
default is 0 that means every sample must be considered}

\item{regions_ordering}{list of ORDER objects where every object contains 
the name of region schema value.
The ORDER's available are: ASC, DESC.
Every condition accepts only one string value. (e.g. DESC("pvalue") )}

\item{rtop}{integer value specifying the first k regions in each group.
default is 0 that means every sample must be considered}

\item{rtopg}{integer value specifying the first k regions to retrieve 
in each group.
default is 0 that means every sample must be considered}

\item{rtopp}{integer value specifying the percentage of regions to retrieve.
default is 0 that means every sample must be considered}
}
\value{
DataSet class object. It contains the value to use as input 
for the subsequent GMQL function
}
\description{
It is used to order either samples or sample regions or both, according to 
a set of metadata and/or region attributes, and/or region coordinates.
Order can be specified as ascending / descending for every attribute
The number of samples and their regions remain the same 
(unless mtop/rtop parameters specified) but a new ordering metadata 
and/or region attribute is added.
Sorted samples or regions have a new attribute "order", 
added to either metadata, or regions, or both of them as specified in input
The input mtop = k and rtop = m extracts the first k samples 
and m regions respectively, the clause mtopg = k and rtopg = m 
performs grouping operation, grouping by identical values 
of ordering attributes and then selects the first k samples 
or regions of each group
}
\details{
mtop, mtopg,mtopp, rtop, rtopg and rtopp are normally numbers: 
if you specify a vector, only the first element will be used
mtop and mtopg and mtopp are mutalbe exclusive, so rtop and rtopg and rtopp
}
\examples{

## It orders the samples according to the Region_count metadata attribute 
## and takes the two samples that have the highest count. 

init_gmql()
test_path <- system.file("example","DATASET",package = "RGMQL")
r = read_dataset(test_path)
o = order(r,list(DESC("Region_Count")), mtop = 2)

}
