% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{ORDERING}
\alias{ORDERING}
\alias{DESC}
\alias{ORDERING}
\alias{ASC}
\title{Ordering functions}
\usage{
DESC(...)

ASC(...)
}
\arguments{
\item{...}{Additional arguments for use in specific methods.}
}
\value{
ordering object
}
\description{
These functions is used to create a series of metadata as string
that require ordering on value.
}
\details{
\itemize{
\item{ASC: It defines a ascending order for input value}
\item{DESC: It defines a descending order for input value}
}
}
\examples{

init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
r = read_dataset(test_path)

## It orders the samples according to the Region_count metadata attribute 
## and takes the two samples that have the highest count. 

desc = arrange(r,list(ASC("Region_Count")), fetch_opt = "mtop", 
num_fetch = 2)
 
asc = arrange(r, list(ASC("Region_Count")),list(DESC("score")),
fetch_opt = "mtop", num_fetch = 5, reg_fetch_opt = "rtop", 
reg_num_fetch = 7)

}
