% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-class.R
\name{Q1}
\alias{Q1}
\title{AGGREGATES object class constructor}
\usage{
Q1(value)
}
\arguments{
\item{value}{string identifying name of metadata attribute}
}
\value{
Q1 aggregate object
}
\description{
This class constructor is used to create instances of AGGREGATES object,
to be used in GMQL functions that require aggregate on value.
It prepares input parameter to be passed to the library function 
first quartile, performing all the type conversions needed
}
\examples{

### local with CustomParser
init_gmql()
test_path <- system.file("example", "DATASET", package = "RGMQL")
exp = read_dataset(test_path)

## This statement copies all samples of exp into res dataset, 
## and then calculates new metadata attributes for each of them: 
## q1_score is the first quartile of score of the sample regions.

res = extend(input_data = exp, list(q1_score = Q1("score")))


}
\seealso{
\code{\link{SUM}} \code{\link{COUNT}} \code{\link{MIN}} 
\code{\link{MAX}} \code{\link{AVG}} \code{\link{MEDIAN}} 
\code{\link{STD}} \code{\link{BAG}} \code{\link{BAGD}}
\code{\link{Q2}} \code{\link{Q3}}
}
