% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{Q3}
\alias{Q3}
\title{OPERATOR object class constructor}
\usage{
Q3(value)
}
\arguments{
\item{value}{single string identifying name of region attribute}
}
\value{
Q3 Operator object
}
\description{
This class constructor is used to create instances of OPERATOR object
to be used in GMQL functions that require operator on value,
it prepared input parameter to be passed to library function third quartile,
performing all the type conversion needed
}
\examples{

### local with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
exp = readDataset(test_path)

## This statement copies all samples of exp into res dataset, and then calculates new 
## metadata attributes for each of them: q3_score is the third quartile of score of the sample regions.
res = extend(input_data = exp, list(q3_score = Q3("score")))

}
\seealso{
\code{\link{MIN}} \code{\link{MAX}} \code{\link{AVG}} \code{\link{COUNT}}
\code{\link{BAG}} \code{\link{STD}} \code{\link{MEDIAN}} \code{\link{Q1}} \code{\link{SUM}}
\code{\link{SUM}}
}
