% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(samples)
}
\arguments{
\item{samples}{GrangesList}
}
\value{
DAGgraph class object. It contains the value associated to the graph used 
as input for the subsequent GMQL function
}
\description{
Read a GrangesList saving in scala memory that can be referenced in R
}
\examples{
"prova prova"

}
