% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{readDataset}
\alias{readDataset}
\title{GMQL Function: READ}
\usage{
readDataset(dataset, parser = "CustomParser", is_local = TRUE, url = NULL,
  override = FALSE)
}
\arguments{
\item{parser}{single string used to parsing dataset files
The Parser's available are:
\itemize{
\item{BedParser}
\item{ANNParser}
\item{BroadProjParser}
\item{BedParser}
\item{NarrowPeakParser}
\item{RnaSeqParser}
\item{CustomParser.}
}
Default is CustomParser.}

\item{is_local}{logical value indicating if dataset is a local or remote dataset
if the remote processing is off you cannot use a remote repository (error occured)}

\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{override}{single logical value}

\item{DatasetFolder}{single string folder path for GMQL dataset or datasetname on repository}
}
\value{
"url-like" string to dataset
}
\description{
Read a GMQL dataset or any other folder containig some homogenus sample
from disk, saving in Scala memory that can be referenced in R
Also used to read a repository dataset.
}
\details{
Normally a GMQL dataset contains an XML schema file that contains
name of column header. (e.g chr, start, stop, strand)
The CustomParser read this XML schema; if you already know what kind of schema your files are use one of
the parser defined without reading any XML schema
}
\examples{

\dontrun{

### local with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)

### local with other Parser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path,"ANNParser")
}
""
}
