% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{register_gmql}
\alias{register_gmql}
\title{Register into remote GMQL}
\usage{
register_gmql(url, username, psw, email, first_name, last_name)
}
\arguments{
\item{url}{string url of server: It must contains the server address 
and base url; service name is added automatically}

\item{username}{string user name used to login in}

\item{psw}{string password used to login in}

\item{email}{string user email}

\item{first_name}{string user first name}

\item{last_name}{string user last name}
}
\value{
None
}
\description{
Register to GMQL REST services suite
using the proper GMQL web service available on a remote server.
}
\details{
After registration you will receive an authentication token.
As token remains valid on server (until the next login / registration or 
logout), a user can safely use a token for a previous session as a 
convenience; this token is saved in R Global environment to perform 
subsequent REST calls or batch processing even on complete R restart 
(if the environment has been saved). If error occurs, a specific error 
is printed.
}
\examples{

## Register to GMQL REST services suite 

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
\dontrun{
register_gmql(remote_url,"foo","foo","foo@foo.com","foo","foo")
}

}
