% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_init.R
\name{remote_processing}
\alias{remote_processing}
\title{Disable or Enable remote processing}
\usage{
remote_processing(is_remote)
}
\arguments{
\item{is_remote}{logical value used in order to set the processing mode.
TRUE you set a remote query processing mode, otherwise it will be local,}
}
\value{
None
}
\description{
It allows to enable or disable remote processing
}
\details{
The invocation of this function allows to change mode of processing.
After invoking collect() it is not possbile to switch the processing mode.
}
\examples{

## These statements initializes GMQL with local processing with sample files 
## output format as tab delimited and then change processing mode to remote

init_gmql("tab", remote_processing = FALSE)

remote_processing(TRUE)

}
