% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_read.R
\name{remote_processing}
\alias{remote_processing}
\title{Disable or Enable remote processing}
\usage{
remote_processing(is_remote)
}
\arguments{
\item{is_remote}{logical value used in order to set the processing mode.
TRUE you will set a remote query processing mode otherwise will be local,}
}
\value{
None
}
\description{
It allows to enable or disable remote processing
}
\details{
The invocation of this function allow to change mode of processing.
after invoking collect() is not possbile to switch the processing mode,
}
\examples{

# initialize with remote processing off
init_gmql("tab",remote_processing = FALSE)

# change processing mode to remote
remote_processing(TRUE)

}
