% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{run_query}
\alias{run_query}
\title{Run GMQL query}
\usage{
run_query(url, fileName, query, output_gtf = TRUE)
}
\arguments{
\item{url}{url of server: It must contain the server address 
and base url; service name is added automatically}

\item{fileName}{name of the file}

\item{query}{text of the query}

\item{output_gtf}{logical value indicating file format used for 
storing samples generated by the query.
The possiblities are: 
\itemize{
\item{GTF}
\item{TAB}
}}
}
\value{
None
}
\description{
It runs a GMQL query as single string
}
\details{
If error occures a specific error is printed
}
\examples{

remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
run_query(remote_url, "query_1", "DATASET = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATASET INTO RESULT_DS;", output_gtf = FALSE)

}
