% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{sample_metadata}
\alias{sample_metadata}
\title{Shows metadata list from dataset sample}
\usage{
sample_metadata(url, datasetName, sampleName)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{datasetName}{string name of dataset to get}

\item{sampleName}{string sample name to get}
}
\value{
list of metadata in the form 'key = value'
}
\description{
It retrieves metadata for a specific sample in dataset using the proper 
GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

## download metadata with real test login
remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
sample_metadata(remote_url, "public.HG19_BED_ANNOTATION", "genes")

}
