% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{sample_region}
\alias{sample_region}
\title{Show regions data from a dataset sample}
\usage{
sample_region(url, datasetName, sampleName)
}
\arguments{
\item{url}{string url of server. It must contain the server address
and base url; service name is added automatically}

\item{datasetName}{string name of dataset of interest}

\item{sampleName}{string name of sample of interest}
}
\value{
GRanges data containing regions of sample
}
\description{
It retrieves regions data of a specific sample (whose name is specified in 
the parameter "sampleName") in a specific dataset (whose name is specified 
in the parameter "datasetName") using the proper GMQL web service 
available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

\dontrun{

## Login to GMQL REST services suite as guest

remote_url = "http://www.gmql.eu/gmql-rest/"
login_gmql(remote_url)

## This statement retrieves regions data of sample "S_00000" from public 
## dataset "Example_Dataset_1"
 
sample_region(remote_url, "public.Example_Dataset_1", "S_00000")

}

}
