% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{sample_region}
\alias{sample_region}
\title{Shows regions from a dataset sample}
\usage{
sample_region(url, datasetName, sampleName)
}
\arguments{
\item{url}{string url of server: it must contain the server address
and base url; service name is added automatically}

\item{datasetName}{string name of dataset to get}

\item{sampleName}{string sample name to get}
}
\value{
Granges data containing regions of sample
}
\description{
It retrieves regions for a specific sample 
(whose name is specified in the paramter "sampleName")
in a specific dataset (whose name is specified in the paramter "datasetName") 
using the proper GMQL web service available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{


remote_url = "http://130.186.13.219/gmql-rest"
login_gmql(remote_url)
sample_region(remote_url, "public.HG19_BED_ANNOTATION", "genes")


}
