% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseOp.R
\name{save_query}
\alias{save_query}
\title{Save GMQL query}
\usage{
save_query(url, queryName, queryTxt)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{queryName}{string name of query}

\item{queryTxt}{string text of GMQL query}
}
\value{
None
}
\description{
It saves the GMQL query into repository
using the proper GMQL web service available on a remote server
}
\details{
if you save a query with the same name of an other query already stored 
in repository you will overwrite it,
if no error occures print "Saved" otherwise print the content error
}
\examples{
remote_url = "http://130.186.13.219/gmql-rest"
\dontrun{

login_gmql(remote_url)
save_query(remote_url, "dna_query", "DATASET = SELECT() HG19_TCGA_dnaseq; 
MATERIALIZE DATASET INTO RESULT_DS;")
}
}
\seealso{
\code{\link{show_queries_list}} \code{\link{save_query_fromfile}}
}
