% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{semijoin}
\alias{semijoin}
\title{Semijoin Condtion}
\usage{
semijoin(data, not_in = FALSE, ...)
}
\arguments{
\item{data}{GMQLDataset class object}

\item{not_in}{logical value: T => semijoin is perfomed 
considering semi_join NOT IN semi_join_dataset, F => semijoin is performed 
considering semi_join IN semi_join_dataset}

\item{...}{Additional arguments for use in specific methods.

This method accept a function to define condition evaluation on metadata.
\itemize{
\item{\code{\link{FN}}: Fullname evaluation, two attributes match 
if they both end with value and, if they have a further prefixes,
the two prefix sequence are identical}
\item{\code{\link{EX}}: Exact evaluation, only attributes exactly 
as value will match; no further prefixes are allowed. }
\item{\code{\link{DF}}: Default evaluation, the two attributes match 
if both end with value.}
}}
}
\value{
semijoin condition as matrix
}
\description{
This function is use as support to filter method to define 
semijoin conditions on metadata
}
