% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_metadata.R
\name{show_all_metadata}
\alias{show_all_metadata}
\title{Create a metadata logical map}
\usage{
show_all_metadata(dataset_path)
}
\arguments{
\item{dataset_path}{string with GMQL dataset folder path}
}
\value{
Dataframe
}
\description{
It creates a GRangesList from GMQL samples in dataset. 
It reads sample files in GTF or GDM/tab-delimited format.
}
\examples{

## This statement defines the path to the subdirectory "example" of the 
## package "RGMQL" and imports as GRangesList the contained GMQL dataset

test_path <- system.file("example", "DATASET", package = "RGMQL")
grl = import_gmql(test_path, TRUE)


}
\seealso{
\code{\link{show_all_metadata}}
}
