% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{show_dataset}
\alias{show_dataset}
\alias{show_datasets_list}
\title{Show datasets}
\usage{
show_datasets_list(url)
}
\arguments{
\item{url}{single string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
List of datasets. Every dataset in the list is identified by:
\itemize{
\item{name: name of dataset}
\item{owner: public or name of the user}
}
}
\description{
It shows all GMQL datasets stored in remote repository using the proper 
GMQL web service available on a remote server
}
\details{
If error occurs, a specific error is printed
}
\examples{

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r"
login_gmql(remote_url)
list <- show_datasets_list(remote_url)
}
