% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{show_jobs_list}
\alias{show_jobs_list}
\title{Show all jobs}
\usage{
show_jobs_list(url)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
list of jobs
Every job in the list is identified by:
\itemize{
\item{id: unique job identifier}
}
}
\description{
It show all Jobs (run, succeded or failed) invoked by user
}
\details{
If error occures a specific error is printed
}
\examples{
remote_url = "http://genomic.elet.polimi.it/gmql-rest-r"
login_gmql(remote_url)
list_jobs <- show_jobs_list(remote_url)

}
