% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseOp.R
\name{show_queries_list}
\alias{show_queries_list}
\title{Shows all Queries}
\usage{
show_queries_list(url)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}
}
\value{
list of queries
Every query in the list is identified by:
\itemize{
\item{name: name of query}
\item{text: text of GMQL query}
}
}
\description{
It shows all the GMQL query saved on repository 
using the proper GMQL web service available on a remote server
}
\details{
if error occures, a specific error is printed
}
\examples{

remote_url = "http://130.186.13.219/gmql-rest"

login_gmql(remote_url)
list <- show_queries_list(remote_url)

}
\seealso{
\code{\link{save_query}}
}
