% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{showDatasets}
\alias{showDatasets}
\title{Show Dataset}
\usage{
showDatasets(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name is added automatically}
}
\value{
list of datasets.

Every dataset in the list is identified by:
\itemize{
\item{name: name of dataset}
\item{owner: public or name of the user}
}
}
\description{
It show all GMQL dataset stored in repository using the proper GMQL web service 
available on a remote server
}
\details{
If error occures a specific error is printed
}
\examples{

#### show dataset when logged as guest
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl)
list <- showDatasets(PolimiUrl)

}
\seealso{
\code{\link{deleteDataset}}
}
