% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{showSamplesFromDataset}
\alias{showSamplesFromDataset}
\title{Show dataset samples}
\usage{
showSamplesFromDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get
if the dataset is a public dataset, we have to add "public." as prefix, as shown in the example below
otherwise no prefix needed}
}
\value{
list of samples in dataset.

Every sample in the list is identified by:
\itemize{
\item{id: id of sample}
\item{name: name of sample}
\item{path: sample repository path}
}
}
\description{
It show all sample from a specific GMQL dataset
}
\details{
If error occured a specific error will be printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl)
list <- showSamplesFromDataset(PolimiUrl,"public.GRCh38_ENCODE_BROAD_MAY_2017")

}
\seealso{
\code{\link{uploadSamples}}
}
