% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{STD}
\alias{STD}
\title{OPERATOR object class constructor}
\usage{
STD(value)
}
\arguments{
\item{value}{single string identifying name of region attribute}
}
\value{
STD Operator object
}
\description{
This class constructor is used to create instances of OPERATOR object
to be used in GMQL functions that require operator on value,
it prepared input parameter to be passed to library function standard deviation,
performing all the type conversion needed
}
\examples{

### local with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
exp = readDataset(test_path)

### This statement copies all samples of exp into res dataset, and then calculates new 
metadata attributes for each of them: std_score is the standard deviation score of the sample regions.
res = extend(input_data = exp, list(std_score = STD(score))

}
\seealso{
\code{\link{MIN}} \code{\link{MAX}} \code{\link{AVG}} \code{\link{COUNT}}
\code{\link{BAG}} \code{\link{SUM}} \code{\link{MEDIAN}} \code{\link{Q1}} \code{\link{Q2}}
\code{\link{Q3}}
}
