% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql_init.R
\name{stop_gmql}
\alias{stop_gmql}
\title{Stop GMQL server}
\usage{
stop_gmql()
}
\value{
None
}
\description{
Stop GMQL server
}
\examples{

## These statements initializes GMQL with local processing with sample 
## files output format as tab delimited and then stop it

init_gmql("tab", FALSE)

stop_gmql()

}
