% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{stop_job}
\alias{stop_job}
\title{Stop a job}
\usage{
stop_job(url, job_id)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{job_id}{string id of the job}
}
\value{
None
}
\description{
It stops a specific current query job
}
\details{
If error occurs, a specific error is printed
}
\examples{

\dontrun{

## Login to GMQL REST services suite at remote url

remote_url = "http://genomic.deib.polimi.it/gmql-rest-r/"
login_gmql(remote_url)

## This statement shows all jobs at GMQL remote system and selects one 
## running job, saving it into 'jobs_1' (in this case is the first of the 
## list), and then stop it

list_jobs <- show_jobs_list(remote_url)
jobs_1 <- list_jobs$jobs[[1]]
stop_job(remote_url, jobs_1)
}

}
