% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-services.R
\name{upload_dataset}
\alias{upload_dataset}
\title{Upload dataset}
\usage{
upload_dataset(url, datasetName, folderPath, schemaName = NULL)
}
\arguments{
\item{url}{string url of server: It must contain the server address 
and base url; service name is added automatically}

\item{datasetName}{name of dataset to create in repository}

\item{folderPath}{string local path to the folder containing the samples 
files}

\item{schemaName}{string name of schema used to parse the samples;
schemaName available are:
\itemize{
\item{NARROWPEAK}
\item{BROADPEAK}
\item{VCF}
\item{BED}
\item{BEDGRAPH}
}
if schemaName is NULL, it looks for a XML schema file to read in the 
folderPath}

\item{isGMQL}{logical value indicating whether it is uploaded a GMQL 
dataset or not}
}
\value{
None
}
\description{
It uploads a folder (GMQL or not) containing sample files using 
the proper GMQL web service available on a remote server: 
a new dataset is created on remote repository
}
\details{
If no error occurs, it prints "Upload Complete", otherwise a specific error 
is printed
}
\examples{

\dontrun{

## This statement defines the path to the folder "DATASET_GDM" in the 
## subdirectory "example" of the package "RGMQL"

test_path <- system.file("example", "DATASET_GDM", package = "RGMQL")

## Login to GMQL REST services suite at remote url

remote_url <- "http://www.gmql.eu/gmql-rest/"
login_gmql(remote_url)

## Upload of GMQL dataset with "dataset1" as name, without specifying any 
## schema 

upload_dataset(remote_url, "dataset1", folderPath = test_path)

}
}
